% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{asExprsMatrix}
\alias{asExprsMatrix}
\title{Convert data table with ID column to matrix with ID as row names}
\usage{
asExprsMatrix(df)
}
\description{
Convert data table with ID column to matrix with ID as row names
}
\keyword{internal}
