% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Shiny module server functions for browsing metadata}
\usage{
browse(input, output, session, dt_id = cdata, dt_var = metadata,
  infoRmd1 = system.file("help/ID_list.Rmd", package = "DIVE"),
  infoRmd2 = system.file("help/browse_data.Rmd", package = "DIVE"))
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{dt_id}{Table data by ID.}

\item{dt_var}{Table data by var.}

\item{inforRmd1}{Helpfile for uploading an ID list.}

\item{inforRmd2}{Helpfile for how browsing data works.}
}
\description{
Handles general browsing, filtering, subsetting of table data
}
