% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdata}
\alias{cdata}
\title{Experimental nPOD data curated from over 70 publications}
\format{A data.table with 460 rows and 218 columns}
\usage{
cdata
}
\description{
A dataset of nPOD cases by variable. This is used by \code{\link{matrixCtrl}}.
}
\keyword{datasets}
