% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPack.R
\name{cellPack}
\alias{cellPack}
\title{Server module function for cellPack widget}
\usage{
cellPack(input, output, session, json, css = system.file("www/",
  "cellpack.css", package = "DIVE"))
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{json}{Relative path to the JSON data for widget.}

\item{css}{Optional, relative path to a CSS file.}
}
\value{
reactive object for element that was clicked
}
\description{
Server module function for cellPack widget
}
