% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPackR.R
\name{cellPackR}
\alias{cellPackR}
\title{Create a custom cell pack widget based on d3 circlepack.js}
\usage{
cellPackR(
  tab,
  entity = "CellTissue",
  context = "CellTissueContext",
  nested = "|",
  cl = system.file("ontology/cl_2018-11.obo", package = "DIVE"),
  rel_type = "is_a",
  ancestor_classes,
  class_colors,
  corrections = NULL,
  contextcolors
)
}
\arguments{
\item{tab}{Table with data for widget.}

\item{entity}{Name of column containing relevant entity.}

\item{context}{Name of column containing context.}

\item{nested}{Deliminater for nested rows, defaulting to "|".}
}
\description{
A wrapper to produce an html widget representing cells and tissues as circles placed in hierarchy.
}
