% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkDataUpload}
\alias{checkDataUpload}
\title{General check for uploaded data}
\usage{
checkDataUpload(data, result = NULL, idcol = "ID", idchar = TRUE, html = TRUE)
}
\arguments{
\item{data}{A data.table.}

\item{result}{What to return in place of data if issues are found with the data.}

\item{idcol}{Name of ID column, defaulting to "ID".}

\item{idchar}{Logical with default TRUE, whether to convert ID column to character.}

\item{html}{Logical with default TRUE, whether to format message for HTML display.}
}
\value{
A list with  \code{result} and  \code{message}.
}
\description{
This is generally meant for dataset upload modules. Checks for (in this order):
\enumerate{
  \item one ID column, which is by default converted to character
  \item actual data columns besides ID
  \item all data columns are numeric
}
If none of the above, returns the data in \code{result} and NULL in \code{message},
otherwise the method witholds the data by returning NULL and
message will be non-NULL, containing the issue noted.
For some use cases might want return data even if there are issues
by specifying \code{result = data}.
}
