% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkForID}
\alias{checkForID}
\title{Check and make character ID column}
\usage{
checkForID(data, message = NULL, idcol = "ID", aschar = T)
}
\arguments{
\item{data}{A table.}

\item{idcol}{Name of ID column, defaulting to "ID".}

\item{aschar}{Convert ID column to character.}
}
\value{
A list with elements result with the result data and message with a message to show.
}
\description{
Checks for some sort of ID column, which is by default converted to character.
Returns a message for failures.
This is meant to work with dataset upload modules.
}
