% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO.R
\name{checkGEOmatrixfile}
\alias{checkGEOmatrixfile}
\title{Check the GEO series matrix file}
\usage{
checkGEOmatrixfile(eset)
}
\description{
Check the GEO series matrix file
}
\details{
Accessions may not contain data in the matrix file but in the supplement section as an Excel file,
which means data cannot by imported.
}
\seealso{
Other checkGEO functions: 
\code{\link{checkGEOmetatype}()},
\code{\link{checkGEOplatform}()},
\code{\link{checkGEOwrapper}()}
}
\concept{checkGEO functions}
\keyword{internal}
