% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO.R
\name{checkGEOmetatype}
\alias{checkGEOmetatype}
\title{Check GEO metadata for supported dataset type}
\usage{
checkGEOmetatype(
  eset,
  supported = c("Expression profiling by array",
    "Expression profiling by high throughput sequencing",
    "Protein profiling by protein array", "Methylation profiling by array",
    "Methylation profiling by high throughput sequencing")
)
}
\description{
Check GEO metadata for supported dataset type
}
\details{
For reference: https://www.ncbi.nlm.nih.gov/gds/advanced/ -> select field "Dataset Type" -> show index list
}
\seealso{
Other checkGEO functions: 
\code{\link{checkGEOmatrixfile}()},
\code{\link{checkGEOplatform}()},
\code{\link{checkGEOwrapper}()}
}
\concept{checkGEO functions}
