% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO.R
\name{checkGEOplatform}
\alias{checkGEOplatform}
\title{Check the GEO platform}
\usage{
checkGEOplatform(eset)
}
\description{
Check the GEO platform
}
\details{
Within the broad dataset types, handled by \code{\link{checkGEOmetatype}},
the different types of platforms that need to be considered for compatibility.
This excludes older platforms that are not one-channel array
for the "Expression profiling by array" type by using information in \code{pData},
though there might be a better way to do this check.
}
\seealso{
Other checkGEO functions: 
\code{\link{checkGEOmatrixfile}()},
\code{\link{checkGEOmetatype}()},
\code{\link{checkGEOwrapper}()}
}
\concept{checkGEO functions}
\keyword{internal}
