% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO.R
\name{checkGEOwrapper}
\alias{checkGEOwrapper}
\title{Check a GEO dataset for compatibility}
\usage{
checkGEOwrapper(eset)
}
\description{
Check a GEO dataset for compatibility
}
\details{
Primarily used in \code{\link{getGEOServer}}, this wraps multiple checks of varying complexity
that were compartmentalized to help iteration of adaptable and reliable check steps over time.
}
\seealso{
Other checkGEO functions: 
\code{\link{checkGEOmatrixfile}()},
\code{\link{checkGEOmetatype}()},
\code{\link{checkGEOplatform}()}
}
\concept{checkGEO functions}
\keyword{internal}
