% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFromCollection.R
\name{checkMeta}
\alias{checkMeta}
\title{Check metadata for \preformatted{cdata}}
\usage{
checkMeta(cdata, metadata)
}
\arguments{
\item{cdata}{A data.table, typically output of cdataMake.}

\item{metadata}{File containing metadata.}
}
\value{
A summary that includes which features are missing what type of metadata.
}
\description{
Checks that data has available matching metadata.
}
