% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPlot.R
\name{cohistPlot}
\alias{cohistPlot}
\title{A custom comparison histodot plot}
\usage{
cohistPlot(data, x = "Cohort", y, fill = "Matched")
}
\description{
A custom comparison histodot plot
}
\keyword{internal}
