% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortGraph.R
\name{cohortGraph}
\alias{cohortGraph}
\title{Shiny module server for interactive cohort graph}
\usage{
cohortGraph(input, output, session, cohgraph, cohdata)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{cohgraph}{Pre-built plotly graph widget.}

\item{cohdata}{Accompanying data.table of cohort data graph was built off.}
}
\value{
A hybrid parallel axes graph for exploring a dataset (typically cohort).
}
\description{
Shiny module server for interactive cohort graph
}
