% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortGraph.R
\name{cohortGraphOutput}
\alias{cohortGraphOutput}
\title{Shiny UI Output for interactive cohort graph}
\usage{
cohortGraphOutput(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\value{
A \preformatted{div} for the plots.
}
\description{
The module embeds an interactive "hybrid parallel coordinates" graph pre-built,
linked with a pie chart that gives more information on a node hover.
}
