% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{colAgg}
\alias{colAgg}
\title{Coalesce-like function for combining rows}
\usage{
colAgg(x)
}
\arguments{
\item{x}{A `data.table`.}
}
\description{
Example:
ID   var1 var2   var3  var4   var5
001      0.313      0.344   NA     NA     NA
001         NA         NA 22.4 41.111 20.669
becomes -->
ID var1 var2   var3  var4   var5
001      0.313      0.344 22.4 41.111 20.669
}
\keyword{internal}
