% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{collectData}
\alias{collectData}
\title{Shiny module server for data collection}
\usage{
collectData(input, output, session, template = reactive({ }),
  filldata = reactive({ }), class = "")
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{template}{UI template.}

\item{filldata}{A list of existing data values that can be passed in. If NULL, works as if in "add" instead of "edit" mode.}

\item{class}{Optional, class for ui CSS.}
}
\description{
Generates UI according to a template and returns input data. Fields can be pre-filled with existing data.
}
