% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{collectDataInput}
\alias{collectDataInput}
\title{Shiny module UI for data collection}
\usage{
collectDataInput(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
UI contains input fields to collect data.
}
