% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPackR.R
\name{colorByClass}
\alias{colorByClass}
\title{Color By Class}
\usage{
colorByClass(dt, e_ancestry, indexed, colors = NULL)
}
\arguments{
\item{dt}{A data.table with entities for which a color column will be added.}

\item{e_ancestry}{A list of ancestor classes by entity.}

\item{indexed}{A list of the same length as e_ancestry, treated as factors and used for splitting e_ancestry into defined classes.}

\item{colors}{A vector of the same length as unique values in indexed, mapping classes to specific color hues.
If not specified, random colors are assigned; note that because there are only eight different hues available
c("red", "orange", "yellow", "green", "blue", "purple", "pink", "monochrome"),
when there are more than eight classes, some classes will necessarily share the same hue.}
}
\description{
Color By Class
}
