% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPanel.R
\name{colRange}
\alias{colRange}
\title{Return range for a column in a table}
\usage{
colRange(tbl, col)
}
\arguments{
\item{tbl}{Table of data.}

\item{col}{Name of column.}
}
\description{
Return range for a column in a table
}
\details{
This is a convenience function for reporting range, e.g. as the selection of a Shiny range slider component.
}
\keyword{internal}
