% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPanel.R
\name{colUniqueValues}
\alias{colUniqueValues}
\title{Return unique values for a column in a table}
\usage{
colUniqueValues(tbl, col)
}
\description{
Return unique values for a column in a table
}
\details{
This is a convenience function for reporting values/getting the selection of a Shiny input component.
}
\keyword{internal}
