% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{connectByCases}
\alias{connectByCases}
\title{Connect observations by ID}
\usage{
connectByCases(cdir = getwd(), filepattern = "*[.](txt|tsv|csv)$",
  namepattern = "([^\\\\.]*)", connection = "ID")
}
\arguments{
\item{cdir}{Directory containing collection of dataset files. Defaults to working directory.}

\item{filepattern}{Pattern for selecting files to include.}

\item{namepattern}{Pattern for extracting node names from source file name. Defaults to using the file name without the extension.}

\item{connection}{Name of column used for calculating connections; should be present in all datasets. Defaults to "ID".}
}
\description{
Connect observations by ID
}
