% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consortUI.R
\name{consortUI}
\alias{consortUI}
\title{Connected Sort UI}
\usage{
consortUI(
  inputId,
  label,
  items = NULL,
  item_class = NULL,
  classes = NULL,
  placeholder = "n/a",
  width = "300px",
  ...
)
}
\arguments{
\item{label}{Element label.}

\item{items}{Optional, character vector of items.}

\item{item_class}{Optional, additional class for items in the sortable ui element.}

\item{classes}{Optional, additional classes for the sortable ui element.
Note: including the "one-to-one" class will create a "monogamous" consort element that will only take one item.}

\item{placeholder}{Placeholder text.}
}
\description{
Create a CONnected SORT element whose items are sortable and transferable between other elements initiated by this method.
}
\details{
This is based on the package function \code{shinyjqui::orderInput}, where the R input builder is simpler
and the custom Shiny input bindings are implemented differently.
}
