% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastV.R
\name{contrastVServer}
\alias{contrastVServer}
\title{Shiny module server for creating comparison plots between subgrouped data}
\usage{
contrastVServer(id, cdata, selected_dataset)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{cdata}{A \code{data.table} of clinical, phenotype, or other experimental data used to define groups.}

\item{selected_dataset}{An expression dataset.}
}
\description{
Create comparison plots based on user-defined contrast groups
}
\details{
The implementation primarily relies on \code{limma}.
}
\seealso{
Other contrastV module functions: 
\code{\link{contrastVUI}()}
}
\concept{contrastV module functions}
