% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cordata}
\alias{cordata}
\title{Correlation data}
\format{A list containing a matrix M, correlations in \code{cdata}, and matrix N.}
\usage{
cordata
}
\description{
A dataset generated by \code{\link{data2cor}} with \code{cdata} as input.
This is used by \code{\link{matrixCtrl}}.
}
\keyword{datasets}
