% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{cordataMake}
\alias{cordataMake}
\title{Simple wrapper to make cordata}
\usage{
cordataMake(cdata)
}
\arguments{
\item{cdata}{Output from \code{\link{cdataMake}}.}
}
\description{
Simple wrapper to make cordata
}
