% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customDataset.R
\name{customDatasetInput}
\alias{customDatasetInput}
\title{Shiny module UI for customized user-uploaded dataset}
\usage{
customDatasetInput(id, label = NULL)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{label}{Optional, a label for the name text input field.}
}
\description{
An extension of \code{\link{dataUploadUI}}
}
\details{
The UI extends \code{\link{dataUploadUI}} with an additional text field
for a custom data value that can be appended to the uploaded data.
}
\seealso{
Other customDataset functions: 
\code{\link{customDatasetServer}()}
}
\concept{customDataset functions}
