% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customDataset.R
\name{customDatasetServer}
\alias{customDatasetServer}
\title{Shiny module server for customized user-uploaded dataset}
\usage{
customDatasetServer(
  id,
  customdata,
  defaultvalue = NULL,
  checkFun = NULL,
  informd = NULL,
  appdata = NULL
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{customdata}{Optional, a named list containing name/label to be used in creating the key column.}

\item{defaultvalue}{Optional, a default value to use.}

\item{checkFun}{Optional, a custom check function for an additional layer of checking/modifying uploaded data.
It should return a list containing message and result (result should be \code{NULL} when data fails checks).}

\item{informd}{Relative path to the Rmarkdown file, whose contents will be displayed in the modal.}

\item{appdata}{Optional, the name (including extension) of one or more files stored in appdata that can be
mock-uploaded. See details.}
}
\value{
A reactive \code{data.table} of the modified uploaded data.
}
\description{
An extension of \code{\link{dataUploadServer}}
}
\details{
The server extends \code{\link{dataUploadServer}} by appending to the uploaded data
a custom data column with name specified by \code{customdata} and value filled from user input.
This is conceptually the same as described in \code{\link{dataSubsetServer}},
with the only difference being that \code{customdata} is appended to a user-uploaded
dataset instead of an existing subsettable dataset.
}
\seealso{
Other customDataset functions: 
\code{\link{customDatasetInput}()}
}
\concept{customDataset functions}
