% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastV.R
\name{customGroupSelect}
\alias{customGroupSelect}
\title{Return a select input or slider range depending on group selection}
\usage{
customGroupSelect(x, inputId)
}
\description{
Return a select input or slider range depending on group selection
}
\keyword{internal}
