% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchResult.R
\name{dataFusion}
\alias{dataFusion}
\title{Fusing two datasets based on harmonized variable names}
\usage{
dataFusion(d1, d2, fuseon, sourcecol)
}
\arguments{
\item{d1}{A data.frame of the first dataset.}

\item{d2}{A data.frame of the second dataset.}

\item{fuseon}{A named vector of the harmonized features,
where the names are the features in d1 and elements are features in d2.}

\item{sourcecol}{The key column used to identify the row sources after the two datasets are fused.}
}
\value{
A data.table of the fused data.
}
\description{
Given two datasets and information on which columns are the same, the two datasets
are concatenated, with any NAs removed and keeping only the specified columns in the result.
}
