% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper2.R
\name{dataHelper2Server}
\alias{dataHelper2Server}
\title{Shiny module server function for simple search and filtering of relational tables in a database}
\usage{
dataHelper2Server(
  id,
  dbcon = NULL,
  lhdata,
  lhdatakey = "ID",
  lhfilterconf,
  rhdata,
  rhdatakey = "ID",
  rhfilterconf,
  handler
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{dbcon}{A database connection object to the database with the required tables.}

\item{lhdata}{Name of the table on the left-hand side.}

\item{lhdatakey}{Name of unique key column in \code{lhdata} table, defaulting to "ID".}

\item{lhfilterconf}{If given, will create filters for the left-hand dataset using \code{\link{sideFilterUI}}.
Each named list should have name match a column \code{lhdata} and
also named configuration values which are passed to \code{\link{sideFilterUI}}:
\describe{
  \item{\code{input}}{ One-character vector specifying input type. }
  \item{\code{selected}}{ Initial selection type. }
  \item{\code{conditional}}{ Boolean value for whether this is a filter that should initially be shown or hidden. }
}}

\item{rhdata}{Name of the table on the right-hand side.}

\item{rhdatakey}{Name of unique key column in \code{rhdata}, defaulting to "ID".}

\item{rhfilterconf}{See \code{lhfilterconf}.}

\item{handler}{Name of table used for translating foreign keys between \code{lhdata} and \code{rhdata}.}
}
\description{
Implement interface to traverse data in "right-hand" and "left-hand" tables with additional attribute filters
}
\details{
The simple interface implemented allows constrained browsing, filtering, subsetting of traditional tabular data.
The module requires these tables:
\describe{
  \item{\code{lhdata}}{ "Left-hand" table of data. Conceptually it is also known as a "dimension table".
  Originally, this was a table of case IDs with columns for attributes such as age and sex. }
  \item{\code{rhdata}}{ "Right-hand" table of data. Conceptually it is also known as a "dimension table".
  Originally, this was a table of studies with columns for attributes such as authors, description and reference.}
  \item{\code{handler}}{ Lookup table to translate relations between the left-hand and right-hand tables.
  See helpful reference for \href{https://www.ibm.com/support/knowledgecenter/SSEP7J_11.1.0/com.ibm.swg.ba.cognos.ug_fm.doc/c_dyn_query_bridge_tables.html}{bridge table}}.
}
}
\seealso{
Other dataHelper2: 
\code{\link{dataHelper2UI}()}
}
\concept{dataHelper2}
