% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper2.R
\name{dataHelper2UI}
\alias{dataHelper2UI}
\title{Shiny app UI for simple search and filtering of relational tables in a database}
\usage{
dataHelper2UI(
  id,
  CSS = system.file("www/", "app.css", package = "DIVE"),
  theme = "paper"
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}

\item{theme}{Optional, name of theme for \code{shinythemes}.}
}
\description{
UI to traverse data in "right-hand" and "left-hand" tables with additional attribute data
}
\seealso{
Other dataHelper2: 
\code{\link{dataHelper2Server}()}
}
\concept{dataHelper2}
