% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{dataHelperServer}
\alias{dataHelperServer}
\title{Shiny module server function for simple interactions between two relational tables}
\usage{
dataHelperServer(
  id,
  lhdata,
  lhdatakey = "ID",
  rhdata,
  rhdatakey = "Source",
  initcolselect = NULL,
  handler,
  filterconf
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{lhdata}{A \code{data.table} of the dataset on the left-hand side.}

\item{lhdatakey}{Name of unique key column in \code{lhdata}, defaulting to "ID".}

\item{rhdata}{A \code{data.table} of the dataset on the right-hand side.}

\item{rhdatakey}{Name of unique key column in \code{rhdata}, defaulting to "Source".}

\item{initcolselect}{Initialize column selection (with a checkbox ui) with a vector of names
matching cols in \code{rhdata} to be shown. If \code{NULL}, column selection ui is not rendered.}

\item{handler}{A \code{data.table} used for translating keys between \code{lhdata} and \code{rhdata}.}

\item{filterconf}{If given, will create filters for the left-hand dataset using \code{\link{sideFilterUI}}.
Should be a list with names matching columns in \code{lhdata} to be used as filters
amd elements "type" with their input widget types, "selected" for the initial selection,
and "conditional" with a boolean value for whether this is a less important filter that should initially be hidden/ignored.}
}
\description{
Implement ability to traverse data in "righthand" and "lefthand" tables with additional attribute filters
}
\details{
This implements a simple interface to alllow general browsing, filtering, subsetting of traditional tabular data,
limited to two tables. The module requires these tables:
The "lefthand" table of something with attributes. In the original context, this was a table
of case IDs with attributes such as age and sex.
The "righthand" table of something with other attributes. In the original context,
this was a table of measurements with descriptions and references.
The lookup table to translate relations (use for joins) between the lefthand and righthand tables.
}
\seealso{
Other dataHelper: 
\code{\link{dataHelperUI}()}
}
\concept{dataHelper}
