% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{dataHelperUI}
\alias{dataHelperUI}
\title{Shiny app UI for simple interactions between two relational tables}
\usage{
dataHelperUI(id, CSS = "app.css", theme = "paper")
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}

\item{theme}{Optional, name of theme for \code{shinythemes}.}
}
\description{
UI to traverse data in "righthand" and "lefthand" tables with additional attribute filters
}
\seealso{
Other dataHelper: 
\code{\link{dataHelperServer}()}
}
\concept{dataHelper}
