% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSubset.R
\name{dataSubsetInput}
\alias{dataSubsetInput}
\title{Shiny module UI for selecting a subsetted dataset}
\usage{
dataSubsetInput(id, label = NULL)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{label}{Optional, a label for the datase;
currently this goes into the placeholder rather than as input label}
}
\value{
A \code{shiny::\link[shiny]{tagList}} for UI to subset a dataset.
}
\description{
Create a selection input that subsets a dataset
}
\seealso{
Other dataSubset functions: 
\code{\link{dataSubsetServer}()}
}
\concept{dataSubset functions}
