% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSubset.R
\name{dataSubsetServer}
\alias{dataSubsetServer}
\title{Shiny server module for returning a subsetted dataset}
\usage{
dataSubsetServer(
  id,
  dataset,
  subsetv,
  subsets = NULL,
  customdata = NULL,
  informd = NULL
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{dataset}{A \code{data.table}.}

\item{subsetv}{Name of the subset-by variable in the data.}

\item{subsets}{Optional, a list to display groups with higher-level grouping in the selection.
By default, the selection is composed of unique values in \code{subsetv}.}

\item{customdata}{Optional, a length-one character vector of data column to append. See details.}

\item{informd}{Relative path to the Rmarkdown file, whose contents will be displayed in the modal.}
}
\value{
A reactive \code{data.table} representing a subset of \code{dataset}.
}
\description{
Create a reactive subsetted dataset based on user selection
}
\details{
This modularizes the usual means of subsetting data by selecting group(s) in a variable.
There can be several of these modules to create subsets from the \emph{same} dataset,
e.g. two modules with different grouping variables given to \code{subsetv},
or to create subsets from different datasets.
The only potentially interesting extra option is \code{customdata},
which appends a column to the subsetted result and exists for different purposes.
When there are several instances of the module for creating subsets from different datasets,
appending a \code{customdata} column could be useful to track the dataset origin of the subset,
especially if the subset data are to be merged with other data into one dataset later on.
For example, for two datasets that come from different sources,
\code{customdata} can be passed in as \code{c(Source="S1")} for the first module
and \code{c(Source="S2")} for the second.
Alternatively, \code{customdata} can be useful for appending custom data
such as \code{c(Color="red")}, perhaps to use in plotting of the subsetted data.

Note: currently, multi-selection is a non-configurable default.
}
\seealso{
Other dataSubset functions: 
\code{\link{dataSubsetInput}()}
}
\concept{dataSubset functions}
