% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUpload.R
\name{dataUploadUI}
\alias{dataUploadUI}
\title{Shiny module UI for data upload module}
\usage{
dataUploadUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\value{
UI components.
}
\description{
Generates Shiny UI for the dataUpload module, which contains the main file input
and two optional features including a file-remove button and infolink,
the second of which is intended to be useful for communicating file upload requirements
or instructions.
}
