% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{defaultIndex}
\alias{defaultIndex}
\title{Default index maker}
\usage{
defaultIndex(filenames)
}
\arguments{
\item{filenames}{Character vector of names.}
}
\description{
Create index using file names, e.g. "PMID20062967_Gianani-2010.txt" -> "Gianani10"
}
\keyword{internal}
