% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meh.R
\name{dev_mode}
\alias{dev_mode}
\title{Get and set app dev mode}
\usage{
dev_mode(setting = NULL)
}
\arguments{
\item{setting}{TRUE or FALSE, passed to \code{base::options} \code{DIVE.dev.mode}.}
}
\description{
Control the display of error messages in \code{\link{meh}} and of certain other UI
}
\details{
Aside from \code{\link{meh}}, other modules that use this setting are:
\itemize{
  \item \code{\link{matchResultServer}}: intermediate results are only downloadable in dev mode
}
}
