% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualDrilldown.R
\name{dualDrilldownUI}
\alias{dualDrilldownUI}
\title{Shiny module UI for data drilldown interaction}
\usage{
dualDrilldownUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Creates app UI for a drilldown plot component
}
