% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excludePresent}
\alias{excludePresent}
\title{Exclude values in dataset #2 that are present in dataset #1}
\usage{
excludePresent(x = "ID", data1, data2)
}
\description{
Exclude values in dataset #2 that are present in dataset #1
}
\keyword{internal}
