% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreMore.R
\name{exploreMore}
\alias{exploreMore}
\title{Shiny module server for generating comparative histodot plots of attributes in two datasets}
\usage{
exploreMore(input, output, session, s1Data, s2Data, datakey, refname,
  results, factorx)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{s1Data}{The full "source 1" dataset, considered the reference dataset.}

\item{s2Data}{Reactive "source 2" data that will be compared to the first dataset, usually coming from newDatasetInput.}

\item{results}{The reactive values return of \code{matchResult}.}

\item{factorx}{Attribute name suffix/pattern to indicate which should be plotted as factors.}
}
\value{
Histodot plots for comparing composition and matches between s1Data and s2Data.
}
\description{
Compare the composition between two datasets. In the cohort matching use case, if there is match data,
the plots show which subsets are matched and provides a means of visualizing the match results.
}
