% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreMore.R
\name{exploreMoreUI}
\alias{exploreMoreUI}
\title{Shiny module UI for attribute-by-attribute comparison of two datasets}
\usage{
exploreMoreUI(id, s1Label = "", s1Data, s2Label = "", s2Data, placeholder = "")
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{s1Label}{Optional, name or title of "source 1" dataset to display.}

\item{s1Data}{The full (non-reactive) "source 1" dataset, considered the reference dataset.}

\item{s2Label}{Optional, name or title of "source 2" dataset to display.}

\item{s2Data}{Reactive "source 2" data that will be compared to the first dataset, usually coming from newDatasetInput.}

\item{placeholder}{Optional, placeholder for attribute selection menus.}
}
\value{
Menu UI for selecting attributes from s1 and s2 datasets.
}
\description{
Allows comparison of two datasets (cohorts, in the original case).
}
