% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPackR.R
\name{extendLine}
\alias{extendLine}
\title{Extends the lineage "history" to match global lineage determination.}
\usage{
extendLine(family, extension)
}
\arguments{
\item{family}{List containing values within a conceptual group.}

\item{extension}{A list of possible extensions to prepend to items in family. The longest extension will be used.}
}
\description{
Extends the lineage "history" to match global lineage determination.
}
