% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{filter4j}
\alias{filter4j}
\title{Helper function for extracting inputs created with \code{\link{sideFilterUI}}}
\usage{
filter4j(dt, col, values, j, type)
}
\arguments{
\item{dt}{A \code{data.table}.}

\item{col}{Column in \code{dt}.}
}
\description{
Get values stored in inputs created by \code{\link{sideFilterUI}}
}
