% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper2.R
\name{filterPanel2Server}
\alias{filterPanel2Server}
\title{Filter server}
\usage{
filterPanel2Server(id, dt, dtkey = "ID", filterconf)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{dt}{A database table.}

\item{dtkey}{Name of unique key column in \code{dt}, defaulting to "ID".}

\item{filterconf}{A list with matching columns in \code{dt} to be used as filters
and elements "input" with their input widget types, "selected" for the initial selection,
and "conditional" with a boolean value for whether this is a filter that should initially be hidden/ignored.}
}
\description{
Process and return output from user interaction with corresponding ui
}
\details{
Users are essentially filtering upon values in columns of \code{dt};
the applied combination gives output \code{dtkey} entries,
the calculation of which is actually delegated to the function \code{filter4j}.
Because it may not make sense to use all filters,
\code{filterconf} allows specifying which combination of filters should be active.
}
