% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper2.R
\name{filterPanel2UI}
\alias{filterPanel2UI}
\title{Filter panel UI}
\usage{
filterPanel2UI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
A panel with a combination of filter inputs
}
\details{
Components are actually rendered using the function \code{sideFilterUI}
}
