% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{filterPanelServer}
\alias{filterPanelServer}
\title{Filter server}
\usage{
filterPanelServer(id, dt, dtkey = "ID", filterconf)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{dt}{A \code{data.table}.}

\item{dtkey}{Name of unique key column in \code{dt}, defaulting to "ID".}

\item{filterconf}{A list with names matching columns in \code{lhdata} to be used as filters
and elements "type" with their input widget types, "selected" for the initial selection,
and "conditional" with a boolean value for whether this is a filter that should initially be hidden/ignored.
See \code{\link{sideFilterUI}}.}
}
\description{
Process and return output from user interaction with corresponding ui
}
\details{
Users are essentially filtering upon values in columns of \code{dt};
the applied combination gives output \code{dtkey} entries,
the calculation of which is actually delegated to the function \code{filter4j}.
Because it may not make sense to use all filters,
\code{filterconf} allows specifying which combination of filters should be active.
}
