% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{filterPanelUI}
\alias{filterPanelUI}
\title{Filter panel UI}
\usage{
filterPanelUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
A panel with a combination of filter inputs
}
\details{
Components are actually rendered using the function \code{sideFilterUI}
}
