% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper2.R
\name{filterQuo}
\alias{filterQuo}
\title{Generate filter expression for filter inputs}
\usage{
filterQuo(col, values, type)
}
\arguments{
\item{col}{Filter column in \code{dt}.}

\item{values}{The filter value(s) in `col`.}

\item{type}{Type of input filter; only "range" is special and requires `values` to be a 2-element vector.}
}
\description{
Return a quosure of the appropriate filter expression
for the given input component type to use later on
}
\details{
Most input components serve user selection/filter purposes;
\code{\link{sideFilterUI}} creates these input ui while this function
formulates the selection/filter expression from the input values.
In order to compose the appropriate expression,
we still need some info about the input component given by the parameter \code{type},
e.g. whether it is "range" for a range sliderInput or something else.
}
\keyword{internal}
