% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{functionality}
\alias{functionality}
\title{functionality}
\format{
A data frame with 32 rows and 1 column:
\describe{
\item{Value}{ecosystem performance as defined by model}
}
}
\source{
made-up.
}
\usage{
functionality
}
\description{
measure of ecosystem performance at the same time of sample community
}
\keyword{datasets}
