% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneVUI.R
\name{geneV}
\alias{geneV}
\title{Shiny module server for filtering high dimensional genomics data}
\usage{
geneV(input, output, session, choices, prelist = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{choices}{Choices for selectInput.}

\item{prelist}{Optional, a named list of source files that store pre-compiled sets for convenient access.}
}
\value{
A vector that can be used to subset a high dimensional matrix,
which should be passed in to the parameter \code{selected} in the \code{\link{multiV}} server module.
}
\description{
Currently, this provides options to subset genomics data using custom list input
or GO and Reactome annotations. The output can be passed to multiple "tracks"
in multiVUI for simultaneous filtering.
}
\seealso{
Other multiVApp module functions: 
\code{\link{selectVUI}()},
\code{\link{subgroupVUI}()},
\code{\link{subgroupV}()}
}
\concept{multiVApp module functions}
