% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneV.R
\name{geneVServer}
\alias{geneVServer}
\title{Shiny module server for filtering high dimensional genomics data}
\usage{
geneVServer(id, choices)
}
\arguments{
\item{choices}{Choices for selectInput.}

\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{prelist}{Optional, a named list of source files that store pre-compiled sets for convenient access.}
}
\value{
A vector that can be used to subset a high dimensional matrix,
which should be passed in to the parameter \code{selected} in the \code{\link{multiV}} server module.
}
\description{
Provides options to subset genomics data (using gene accessions)
}
\seealso{
Other multiVApp module functions: 
\code{\link{selectVUI}()},
\code{\link{subgroupVServer}()},
\code{\link{subgroupVUI}()}
}
\concept{multiVApp module functions}
