% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneV.R
\name{geneVUI}
\alias{geneVUI}
\title{Shiny module UI for filtering high dimensional genomics data}
\usage{
geneVUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Provides options to subset genomics data (using gene accessions)
}
\seealso{
Other multiVUI module functions: 
\code{\link{multiVServer}()},
\code{\link{multiVUIR}()},
\code{\link{multiVUI}()},
\code{\link{selectVServer}()}
}
\concept{multiVUI module functions}
