% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_md5}
\alias{get_md5}
\title{Compute MD5 hashes recursively for files matching an extension within a
subdirectory in the current path.}
\usage{
get_md5(extension = "fasta", subdirectory = "")
}
\arguments{
\item{extension}{preferred file extension, default: fasta}

\item{subdirectory}{subdirectory containing the files, within the current}
}
\description{
Compute MD5 hashes recursively for files matching an extension within a
subdirectory in the current path.
}
