% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastV.R
\name{getFilterGroup}
\alias{getFilterGroup}
\title{Apply filter values \code{x} and return a defined group}
\usage{
getFilterGroup(x, refids, filtercol, parentdata)
}
\description{
Convenience method that can handle whether filter values are numeric of character/factor
}
\keyword{internal}
