% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO.R
\name{getGEOInput}
\alias{getGEOInput}
\title{Shiny module UI for importing data from GEO}
\usage{
getGEOInput(id, infoRmd = system.file("help/GEO_module.Rmd", package =
  "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Shiny module UI for importing data from GEO
}
