% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLink.R
\name{getLink}
\alias{getLink}
\title{Shiny module server for retrieving internal data with action buttons/links}
\usage{
getLink(input, output, session, link = T, sources, readfun = readLines,
  ...)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{sources}{A vector of source paths matching the order of labels given in \code{\link{getLinkInput}}.}

\item{readfun}{A function to handle reading source files. Defaults to \code{\link[base]{readLines}}.}

\item{...}{Additional arguments for \preformatted{readfun}.}
}
\value{
Data object.
}
\description{
Inserts action buttons/links that that can be used to 
get data from a specified source.
}
