% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{graphModAes}
\alias{graphModAes}
\title{Modify graph aesthetics based on metadata}
\usage{
graphModAes(aes, metadata, sourcecol = "Source", feature = "color",
  featurecol, mapfun)
}
\arguments{
\item{aes}{Data frame containing nodes to which aesthetics are added or modified.}

\item{metadata}{Either a path to a metadata file or a data.table already in the environment.}

\item{sourcecol}{Column used for matching node names, defaults to "Source".}

\item{feature}{An aesthetic feature, defaults to "color".}

\item{featurecol}{Column used for mapping to the node feature.}

\item{mapfun}{Function that maps values in the feature column to the desired feature labels.}
}
\value{
A data frame containing aesthetics data for the nodes.
}
\description{
Change or add node size, color, or other values depending on metadata.
}
