% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLink.R
\name{guessMatch}
\alias{guessMatch}
\title{Custom function for trying to harmonize diabetes-relevant variables in two different  datasets}
\usage{
guessMatch(extvars)
}
\arguments{
\item{vars}{Vector of variable names from an external (cohort) dataset.
An attempt will be made to harmonize them to variable names in the reference dataset.}
}
\value{
List of named list elements, containing only those reference variables that seem to have
a corresponding version in the external variables (list elements).
}
\description{
Custom function for trying to harmonize diabetes-relevant variables in two different  datasets
}
