% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{hdlistchoicesMake}
\alias{hdlistchoicesMake}
\title{Make custom selection from final hdlist}
\usage{
hdlistchoicesMake(hdlist)
}
\arguments{
\item{hdlist}{Output of \code{\link{hdlistMake}}.}
}
\description{
This should follow \code{\link{hdlistMake}} since
\code{\link{multiVServer}} requires passing in a list object for `choices`.
}
