% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPCGraph.R
\name{HPCGraph}
\alias{HPCGraph}
\title{Shiny module server for interactive hybrid parallel coordinates graph}
\usage{
HPCGraph(input, output, session, hpcg, colors)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{hpcg}{Data object produced by \code{\link{newHPCG}}.}
}
\description{
Shiny module server for interactive hybrid parallel coordinates graph
}
