% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPCGraph.R
\name{HPCGraphOutput}
\alias{HPCGraphOutput}
\title{Shiny UI Output for interactive hybrid parallel coordinates graph}
\usage{
HPCGraphOutput(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\value{
\preformatted{div} component to contain graphs.
}
\description{
The module embeds an interactive "hybrid parallel coordinates graph" built
with \code{\link{hybridParCoordsGraph}}. Categorical nodes are linked to a pie chart
that gives more information on user hover.
}
