% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMatrix.R
\name{iMatrixUI}
\alias{iMatrixUI}
\title{Shiny module UI for interactive matrix plot with drilldown components}
\usage{
iMatrixUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Creates app UI for a heatmap matrix responding to reactive data that is linked to a drilldown scatterplot component.
}
