% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{inCommon}
\alias{inCommon}
\title{Generate an adjacency matrix representing some observation in common for multiple sources}
\usage{
inCommon(cdir = getwd(), filepattern = "*[.](txt|tsv|csv)$",
  namepattern = "([^\\\\.]*)", connection = "ID")
}
\arguments{
\item{cdir}{Directory containing collection of source files. Defaults to working directory.}

\item{filepattern}{Pattern for selecting files to include.}

\item{namepattern}{Pattern for extracting node names from source file name. Defaults to using the file name without the extension.}

\item{connection}{Name of column used for calculating common connection; this column should be present in all datasets. Defaults to "ID".}
}
\value{
An adjacency matrix.
}
\description{
This is a helper for constructing an adjacency matrix based on shared case samples
}
