% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoOutput.R
\name{info}
\alias{info}
\title{Server module function for displaying an info modal}
\usage{
info(input, output, session, infoRmd)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{infoRmd}{Relative path to the Rmarkdown info file, whose contents will be displayed in the modal.}
}
\description{
Server module function for displaying an info modal
}
