% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBox.R
\name{infoBox}
\alias{infoBox}
\title{A UI box element for summary values in dashboard-like presentation}
\usage{
infoBox(title, value, subtitle = NULL, icon = shiny::icon("info"), width = 4)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
This is similar to the UI element in shinydashboard.
}
