% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{infoOutput}
\alias{infoOutput}
\title{Shiny module UI to display a modal on click}
\usage{
infoOutput(id, label = "req's", i = "exclamation-circle", link = T)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{label}{Label, typically for specifying type of info, defaults to "Req's".}

\item{i}{Font-awesome icon name to display next to label, see \code{shiny::\link[shiny]{icon}}.}

\item{link}{Whether to use link or button. Defaults to \code{actionLink}, set FALSE to use \code{actionButton}.}
}
\value{
An actionLink or actionButton that calls a modal.
}
\description{
Create a clickable link or button that displays a modal dialog with customizable content
}
