% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{infoServer}
\alias{infoServer}
\title{Server module function for displaying an info modal}
\usage{
infoServer(id, informd)
}
\arguments{
\item{informd}{Relative path to the Rmarkdown info file, whose contents will be displayed in the modal.}

\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\description{
Server module function for displaying an info modal
}
