% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveMatrix.R
\name{interactiveMatrix}
\alias{interactiveMatrix}
\title{Server module server for plotting correlations matrix with drilldown interaction}
\usage{
interactiveMatrix(input, output, session, state,
  factorx = "grp$|cat$|score$|bin$|count$|pos$",
  dcolors = list(donor.type = c(`Autoab Pos` = "orange", `Cystic
  fibrosis` = "aquamarine4", `Gastric Bypass` = "bisque4", `Gestational
  diabetes` = "deeppink2", `Monogenic Diabetes` = "red4", `No diabetes` =
  "royalblue2", `Other-Diabetes` = "indianred4", `Other-No Diabetes` =
  "steelblue2", T1D = "red", `T1D Medalist` = "maroon", T2D = "purple")))
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{state}{Reactive state variables used for plot functions. See details.}

\item{factorx}{Optional, names and/or patterns for variables that should be displayed as factors in plots.}

\item{dcolors}{List of manual color mappings for variables used for coloring data points.}
}
\description{
The matrix responds to inputs, new plotdata, and has a linked drilldown component.
}
\details{
The plotting logic works with matrixCtrlUI, which returns reactive values required,
including state$plot for the main plot, state$cdata for the drilldown, and state$newdata
storing variable names of new data.
}
