% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveMatrixApp.R
\name{interactiveMatrixApp}
\alias{interactiveMatrixApp}
\title{Shiny app server for exploration of relationships in annotated data with an interactive matrix}
\usage{
interactiveMatrixApp(input, output, session, M = cordata$corM,
  N = cordata$corN, CDATA = cdata, METADATA = metadata,
  widgetopt = "Cell/Tissue", widgetdata = system.file("www/",
  "test.json", package = "DIVE"), infoRmd = "help/data_exploration.Rmd",
  appdata = "pilot.csv")
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{M}{A data matrix, e.g. a correlation matrix, which must have variables as rownames.}

\item{N}{A matrix of the same dimensions as M with data for the filterable layer, e.g. sample size.}

\item{CDATA}{The non-reactive data used for generating the matrix.}

\item{METADATA}{A data.table with "Variable" as a key column and any number of columns (metadata) to be used as filters.}
}
\description{
See matrixCtrlUI
}
