% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveMatrixApp.R
\name{interactiveMatrixAppRun}
\alias{interactiveMatrixAppRun}
\title{Launch Shiny app for exploration of relationships in annotated data with an interactive matrix}
\usage{
interactiveMatrixAppRun()
}
\description{
Wrapper to launch app at console
}
