% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveMatrixApp.R
\name{interactiveMatrixAppUI}
\alias{interactiveMatrixAppUI}
\title{Shiny app UI for exploration of relationships in annotated data with an interactive matrix}
\usage{
interactiveMatrixAppUI(id, CSS = system.file("www/", "app.css", package =
  "DIVE"))
}
\arguments{
\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}
}
\description{
See matrixCtrlUI
}
