% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveMatrix.R
\name{interactiveMatrixUI}
\alias{interactiveMatrixUI}
\title{Shiny module UI of matrix plot with drilldown components}
\usage{
interactiveMatrixUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
The matrix responds to inputs, new plotdata, and has a linked drilldown component.
}
