% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate.R
\name{intermediate}
\alias{intermediate}
\title{Shiny server module for intermediate data processing}
\usage{
intermediate(input, output, session, data, Fun)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{data}{Reactive data.}

\item{Fun}{Function to be applied to data.}
}
\value{
Return value of Fun as reactive object.
}
\description{
This module implements an additional customizable layer for intermediate data processing
by passing data into a function and returning results as a reactive object.
}
