% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate.R
\name{intermediateInput}
\alias{intermediateInput}
\title{Shiny module UI for additional intermediate input}
\usage{
intermediateInput(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\value{
A \code{shiny::\link[shiny]{tagList}} containing input UI.
}
\description{
Adds some UI to provide input for additional data processing
}
