% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFromCollection.R
\name{makeCdata}
\alias{makeCdata}
\title{Make data object for Shiny app}
\usage{
makeCdata(cdir = "./Collection/Main/", filepattern = "*.txt",
  exclude = "!", other = list("./Collection/Other/coredata_ref.txt"),
  outdir = "./App/Data/")
}
\arguments{
\item{other}{Optional, a list of of paths to other data files not in the main collection but
that should be part of the master data object. NULL if no other data to be incorporated.}

\item{outdir}{Where to put \code{cdata}. Defaults to App/Data.}
}
\description{
A wrapper to create the \code{cdata} object
}
