% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchApp}
\alias{matchApp}
\title{Shiny app server for matching and exploration of two datasets}
\usage{
matchApp(input, output, session, refdata, HPCG, colors, refkey, vars,
  guess = DIVE::guessMatch, subsetfeat,
  informd = system.file("help/cohort_exchange.Rmd", package = "DIVE"),
  appdata = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{refdata}{A data matrix, e.g. a correlation matrix, which must have variables as rownames.}

\item{HPCG}{Graph object representing the reference cohort.}

\item{colors}{Colors}

\item{refkey}{Passed to refSubsetInput.}

\item{vars}{Optional, variables allowed to be used for matching within refdata organized in categories.
If not given, all variables in refdata are used without a category.}

\item{guess}{Optional, a function for making initial guesses of matchable variables.}

\item{subsetfeat}{Which variable to be used as the subset variable.}

\item{informd}{Link to to help file}

\item{appdata}{See ?}
}
\description{
Shiny app server for matching and exploration of two datasets
}
\details{
This server function puts together a number of modular server module components with the correct
logic to power the interactive capabilities of the matching application.
}
