% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchAppServer}
\alias{matchAppServer}
\title{Shiny app server for matching and exploration of two datasets}
\usage{
matchAppServer(
  id,
  refdata,
  datakey,
  xname,
  refname,
  vars,
  guess = NULL,
  subsets,
  subsetfeat,
  factorx,
  informd = system.file("help/cohort_exchange.Rmd", package = "DIVE"),
  addon = NULL,
  appdata = NULL
)
}
\arguments{
\item{refdata}{Data.table of reference dataset.}

\item{datakey}{Shared grouping parameter passed to \code{refSubset}, \code{newDataset}, \code{matchResult}.}

\item{xname}{Character label for the external dataset, e.g. "CohortX" or "New Site" or "Comparison Dataset".}

\item{refname}{Character label for the reference dataset.}

\item{vars}{Variables allowed to be used for matching within \code{refdata}, which will be organized and
displayed as the given named lists.}

\item{guess}{Optional, a function for making initial guesses of matchable variables.}

\item{subsets}{The subsets.}

\item{subsetfeat}{Which variable to be used as the subset variable.}

\item{factorx}{A naming pattern for factoring variables displayed.}

\item{informd}{Optional, link to a help file that can be called with the info module.}

\item{addon}{Optional, a self-contained module that doesn't interact with others,
e.g. for extra summary plots. Needs to be passed in as name of the server module function.}

\item{appdata}{Optional, a path to example application data (for demonstration purposes).}

\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\description{
Shiny app server for matching and exploration of two datasets
}
\details{
This server function puts together a number of modular components to
implement the whole interaction of the matching application. Though originally composed for
matching patient cohorts, this matching application can be used with data in different domains.
The composition of this module allows inserting a addon-written module.
}
