% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchAppUI}
\alias{matchAppUI}
\title{Shiny app UI for matching application}
\usage{
matchAppUI(id, subsets, CSS = system.file("www/", "app.css", package =
  "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{subsets}{Optional, a list of the named subset groups. This can be used to allow only certain subsets
within the data available for matching.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}
}
\description{
Shiny app UI for matching application
}
\details{
The application UI can be viewed as a default template of how to put together several module components,
including \code{\link{newDataSetInput}}, \code{\link{refSubsetInput}}, \code{\link{cohortGraphOutput}},
to overall enable interactive exploration and matching of two different (cohort) datasets.
}
