% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLink.R
\name{matchLinkServer}
\alias{matchLinkServer}
\title{Shiny server module for interactive drag-and-drop variable linkage or harmonization}
\usage{
matchLinkServer(
  id,
  refdata,
  inputdata,
  vars,
  guess = NULL,
  informd = system.file("help/matching_methods.Rmd", package = "DIVE")
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{refdata}{Reactive reference \code{data.table}.}

\item{inputdata}{Reactive \code{data.table}, typically provided by \code{\link{newDatasetServer}}.}

\item{vars}{A named list of a variable set (or sets) in the data that is allowed for matching. See details.}

\item{guess}{Optional, name of the function to call for initial guessing of harmonized variables. See details.}

\item{informd}{Optional, relative path to an info Rmarkdown file that will be pulled up in a help modal.
Recommend using for providing methods details of the matching algorithm.}
}
\value{
The returned reactive list is intended to be used as inputs to \code{\link{matchResultServer}}:
\enumerate{
  \item \code{params$matchOpts} contains all the possible variables that could be used between the two datasets
  \item \code{params$matchOn} contains variables that were actually paired up for matching
  \item \code{params$run} transmits the "finalized" params/triggers \code{\link{matchResultServer}}
}
}
\description{
Given two datasets, a "reference" and an "external" or "comparison" dataset,
both presumably containing at least some of the same variables,
this module implements a drag-and-drop interface that allows users to
interactively harmonize variables (pair them) as parameters for matching.
}
\details{
The drops are initialized with the specified \code{vars} in \code{refdata},
and then a "variable bank" is initialized when given \code{inputdata}.
\code{vars} should be provided as a list such as \code{list(Variables = c("age", "sex", "BMI", "biomarkerLevel"))} or
\code{list(Demographic = c("age", "sex"), Clinical = c("BMI", "biomarkerLevel"))}.

A guess function can be provided to attempt guessing which variables are actually the same between the datasets,
which will run and pre-populate the pairs before the user does manual linkage of variables
The guess function should be specific to the dataset and the domain.
For example, if we can expect demographic-type datasets, the function can look for "sex" or "gender"
in the incoming dataset to be matched with "sex" in the reference data,
so that the user doesn't have to specify this manually.
Its return should contain a list with all matchable variables in \code{refdata} and
the match in \code{inputdata}, or \code{NULL} if no match.
}
