% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLink.R
\name{matchLinkUI}
\alias{matchLinkUI}
\title{Shiny module UI for interactive drag-and-drop variable harmonization of two datasets}
\usage{
matchLinkUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\value{
Interactive drag-and-drop UI elements.
}
\description{
Shiny module UI for interactive drag-and-drop variable harmonization of two datasets
}
