% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPlot.R
\name{matchPlotUI}
\alias{matchPlotUI}
\title{Shiny module UI for comparison histodot plots of attributes in two datasets}
\usage{
matchPlotUI(
  id,
  s1label = NULL,
  s2label = NULL,
  placeholder = "(select from attributes)"
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{s1label}{Optional, label for first selection.}

\item{s2label}{Optional, label for second selection.}

\item{placeholder}{Optional, placeholder for selection menus.}
}
\value{
Menu UI for selecting attributes from s1 and s2 datasets.
}
\description{
Histodot plots of two datasets side-by-side, with or without match information
}
\seealso{
Other matchPlot functions: 
\code{\link{matchPlotServer}()}
}
\concept{matchPlot functions}
