% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchResult.R
\name{matchResult}
\alias{matchResult}
\title{Shiny module server for generating match results output}
\usage{
matchResult(input, output, session, refSubset, cohortX, params)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{refSubset}{Reactive reference cohort subset data.table.}

\item{cohortX}{Reactive cohortX data.table, which typically comes from the newCohortInput module.}

\item{params}{Reactive params data, typically from the matchParameters module.}
}
\value{
Reactive values params, intermediate, pair, and matchtable.
}
\description{
Shiny module server for generating match results output
}
