% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchResult.R
\name{matchResultServer}
\alias{matchResultServer}
\title{Shiny module server for generating match results output}
\usage{
matchResultServer(id, refSubset, setX, params, sourcecol)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{refSubset}{Reactive reference cohort subset data.table, i.e. from \code{refSubset}.}

\item{setX}{Reactive data.table of comparison subset, i.e from \code{newDataset} module.}

\item{params}{Reactive parameters data, i.e. from \code{matchLink} module.}

\item{sourcecol}{Name for the source key column of the joined dataset.}
}
\value{
Reactive values containing params, intermediate results, pair, and matchtable. See \code{matchPair}.
}
\description{
Shiny module server for generating match results output
}
