% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixApp.R
\name{matrixApp}
\alias{matrixApp}
\title{Launch Shiny app for exploration of relationships in annotated data with an interactive matrix}
\usage{
matrixApp(ns, ...)
}
\arguments{
\item{ns}{Namespace of app module.}

\item{...}{Arguments passed to \code{\link{matrixAppServer}}.}
}
\description{
Wrapper to launch app at console
}
