% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixApp2.R
\name{matrixApp2}
\alias{matrixApp2}
\title{Launch Shiny app for exploration of relationships in annotated data with an interactive matrix}
\usage{
matrixApp2(ns, ...)
}
\arguments{
\item{ns}{Namespace of app module.}

\item{...}{Arguments passed to \code{\link{matrixApp2Server}}.}
}
\description{
Wrapper to launch app at console
}
