% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixApp2.R
\name{matrixApp2UI}
\alias{matrixApp2UI}
\title{Shiny app UI for interactive matrix (v2)}
\usage{
matrixApp2UI(id, CSS = system.file("www/", "app.css", package = "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}
}
\description{
See matrixCtrl2UI
}
