% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixApp.R
\name{matrixAppServer}
\alias{matrixAppServer}
\title{Shiny app server for exploration of relationships in annotated data with an interactive matrix}
\usage{
matrixAppServer(
  id,
  M,
  N,
  P,
  cdata,
  metadata,
  vkey,
  checkFun = DIVE::checkDataUpload,
  widgetmod = NULL,
  widgetopt = NULL,
  factorx,
  dcolors,
  colorscales = list(default = list(colorscale_named(pal = "RdBu"), zmin = -1, zmax =
    1)),
  informd = system.file("help/interactive_matrix.Rmd", package = "DIVE"),
  appdata = NULL
)
}
\arguments{
\item{M}{A data matrix, e.g. a correlation matrix, which must have variables as rownames.}

\item{N}{A matrix of the same dimensions as M with data for the filterable layer, e.g. sample size.}

\item{P}{A matrix of the same dimensions as M with data for the filterable layer, e.g. p-values.}

\item{cdata}{The non-reactive data used for generating the matrix.}

\item{metadata}{A data.table with "Variable" as a key column and any number of columns (metadata) to be used as filters.}

\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\description{
See matrixCtrlUI
}
