% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixAsNetwork.R
\name{matrixAsNetworkServer}
\alias{matrixAsNetworkServer}
\title{Server module server for network graph}
\usage{
matrixAsNetworkServer(
  id,
  mdata,
  M,
  background = NULL,
  .nodes = NULL,
  .edges = NULL,
  .options = NULL,
  .interaction = list(multiselect = TRUE),
  randomSeed = NULL
)
}
\arguments{
\item{id}{Namespace id for module.}

\item{mdata}{Reactive matrix data from \code{\link{matrixCtrl}}.}

\item{M}{Full matrix from which `mdata` originates.}

\item{style}{Optional, a list with global styling for nodes, edges, and plot background. Defaults to visNetwork styling.}
}
\description{
Making a network graph representation of matrix data using visNetwork.
}
\details{
This is a module with a visNetwork plot as its main output,
meant to implement an alternative view of data as a graph.
As the name of the module suggests, that data should be in matrix form;
the module can use whatever is returned by the matrixCtrl module.
Multi-selection of nodes is set as the default interaction,
so return is the currently selected nodes as a character vector node IDs
(this can be changed to single select).
}
