% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixAsNetwork.R
\name{matrixAsNetworkUI}
\alias{matrixAsNetworkUI}
\title{Shiny module UI for network graph}
\usage{
matrixAsNetworkUI(id, height = "400px", ...)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Create app UI for a network graph
}
\details{
This is a module with a visNetwork plot as its main output,
meant to implement an alternative view of data as a graph.
}
