% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCtrl2.R
\name{matrixCtrl2UI}
\alias{matrixCtrl2UI}
\title{Shiny module UI for matrix view}
\usage{
matrixCtrl2UI(id, minN = 5)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{minN}{Minimum number for numeric input filter.}
}
\description{
Interactive controls of matrix view
}
