% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCtrl.R
\name{matrixCtrlUI}
\alias{matrixCtrlUI}
\title{Shiny module UI containing filter inputs for interactive matrix}
\usage{
matrixCtrlUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Interactive controls for a matrix (usually correlation) plot.
}
