% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixInteractive.R
\name{matrixInteractiveUI}
\alias{matrixInteractiveUI}
\title{Shiny module UI for interactive matrix with drilldown component}
\usage{
matrixInteractiveUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Creates app UI for a heatmap matrix responding to reactive data that is linked to a drilldown scatterplot component.
}
