% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixMain.R
\name{matrixMainUI}
\alias{matrixMainUI}
\title{Shiny module UI for interactive matrix with drilldown component}
\usage{
matrixMainUI(id, ...)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{...}{Other attributes and child elements to pass into UI container.}
}
\description{
Creates app UI for a heatmap matrix responding to reactive data that is linked to a drilldown scatterplot component.
}
