% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixMix.R
\name{matrixMix}
\alias{matrixMix}
\title{Shiny app server for exploration of relationships in annotated data with an interactive matrix}
\usage{
matrixMix(input, output, session, M, N, cdata, metadata, vkey,
  widgetmod = NULL, widgetopt = NULL, factorx, dcolors,
  informd = system.file("help/interactive_matrix.Rmd", package = "DIVE"),
  appdata = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{M}{A data matrix, e.g. a correlation matrix, which must have variables as rownames.}

\item{N}{A matrix of the same dimensions as M with data for the filterable layer, e.g. sample size.}

\item{cdata}{The non-reactive data used for generating the matrix.}

\item{metadata}{A data.table with "Variable" as a key column and any number of columns (metadata) to be used as filters.}
}
\description{
See matrixCtrlUI
}
