% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixMix.R
\name{matrixMixR}
\alias{matrixMixR}
\title{Launch Shiny app for exploration of relationships in annotated data with an interactive matrix}
\usage{
matrixMixR(ns, ...)
}
\description{
Wrapper to launch app at console
}
