% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixMix.R
\name{matrixMixUI}
\alias{matrixMixUI}
\title{Shiny app UI for exploration of relationships in annotated data with an interactive matrix}
\usage{
matrixMixUI(id, CSS = system.file("www/", "app.css", package = "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}
}
\description{
See matrixCtrlUI
}
