% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFromCollection.R
\name{mergeMore}
\alias{mergeMore}
\title{Merge a main data table with other misc data}
\usage{
mergeMore(data, inputpath)
}
\arguments{
\item{data}{A data.table}

\item{inputpath}{Path to another dataset to merge with data.}
}
\value{
A data.table
}
\description{
Use \code{mergeMore} to merge some other data with the master dataset.
}
