% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{metadata}
\alias{metadata}
\title{Metadata for \code{cdata}}
\format{A data.table with 460 rows and 218 columns}
\usage{
metadata()
}
\description{
Contains a table of variables annotated with several types of metadata. 
This \code{metadata} data object is used by \code{\link{matrixCtrl}}
}
