% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVUI.R
\name{multiV}
\alias{multiV}
\title{Shiny app server for multi-views}
\usage{
multiV(input, output, session, hdata = NULL, hcat = NULL,
  cdata = NULL, genes = DIVE::gene_symbols, prelist = NULL,
  slabel = DIVE::gene_symbols_map)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\description{
Assembles the logic of various module components, i.e. \code{\link{multiV}}, \code{\link{geneV}},
into a working one-page application
}
\seealso{
Other multiVApp module functions: \code{\link{geneV}},
  \code{\link{selectVUI}}, \code{\link{selectV}},
  \code{\link{subgroupVUI}}, \code{\link{subgroupV}}
}
\concept{multiVApp module functions}
