% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVApp.R
\name{multiVApp}
\alias{multiVApp}
\title{Shiny app server for multi-views}
\usage{
multiVApp(input, output, session, HDATA = list(`Yip et al. (unpublished)`
  = xm_t, `Liu et al. 2016` = px1_t, `Nyalwidhe et al. 2017` = px2_t),
  CDATA = cdata, CHOICES = gene_symbols)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\description{
Assembles the logic of various module components, i.e. \code{\link{multiV}}, \code{\link{geneV}},
into a working one-page application
}
