% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVApp.R
\name{multiVAppRun}
\alias{multiVAppRun}
\title{Shiny app launcher for multi-view module}
\usage{
multiVAppRun()
}
\description{
Shiny app launcher for multi-view module
}
\seealso{
Other multiVApp module functions: \code{\link{geneVUI}},
  \code{\link{geneV}}, \code{\link{multiVApp}},
  \code{\link{multiVUIApp}}, \code{\link{selectVUI}},
  \code{\link{selectV}}, \code{\link{subgroupVUI}},
  \code{\link{subgroupV}}
}
\concept{multiVApp module functions}
