% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVCtrlUI.R
\name{multiVCtrl}
\alias{multiVCtrl}
\title{Shiny module server for controlling multi-column view}
\usage{
multiVCtrl(input, output, session, hdlist, choices = names(hdlist),
  infoRmd = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{hdlist}{A list containing high dimensional datasets; must have names to be used as choices in the selection menu.}

\item{choices}{Names referencing datasets in hdlist, to be used in selection menu.
When the datasets should be displayed as grouped, the choices can be passed in accordingly for \code{\link[shiny]{selectizeInput}}.}

\item{infoRmd}{Optional link to an Rmarkdown document containing details for the data upload module.}
}
\value{
A list containing the data matrix for the parameter \preformatted{hdata} in the \code{\link{multiV}} module.
}
\description{
Implements sourcing datasets for \code{\link{multiVUI}} modules in the application page.
}
