% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVCtrl.R
\name{multiVCtrlUI}
\alias{multiVCtrlUI}
\title{Shiny module UI for controlling multi-column views}
\usage{
multiVCtrlUI(id, menu = TRUE, upload = TRUE, GEO = TRUE)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{menu}{Logical flag, whether to allow a menu for loading stored datasets.}

\item{upload}{Logical flag, whether to allow data upload.}

\item{GEO}{Logical flag, whether to allow pulling data from GEO (beta).}
}
\description{
Shiny module UI for controlling multi-column views
}
\details{
Multiple high dimensional datasets can be displayed by calling separate instances of
\code{\link{xVUI}} modules, with the contents typically laid out in row containers.
This controller UI has three different ways to select and source the datasets:
\enumerate{
  \item Selecting from available pre-processed datasets.
  \item User-uploaded data.
  \item A beta (least-supported) method of retrieving datasets from GEO.
}
For convenience and customization, it is possible make unavailable any of the three sourcing methods,
e.g. hide the GEO method by not displaying the UI.
}
