% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiV.R
\name{multiVServer}
\alias{multiVServer}
\title{Shiny app server module for multi-views}
\usage{
multiVServer(
  id,
  hdlist = NULL,
  choices = DIVE::hdlistchoicesMake(hdlist),
  cdata = NULL,
  preselect = NULL
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{hdlist}{A list of matrices representing high dimensional datasets; the names are used for \code{choices}.}

\item{choices}{Selection choices are by default created from automatic parsing of `hdlist`. However, a manual list can be given,
which should be appropriate for passing to \code{shiny::\link[shiny]{selectizeInput}}.}

\item{cdata}{A \code{data.table} of characteristics data, commonly phenotype or clinical data.}

\item{preselect}{Optional, pre-selected phenotype or clinical variables from \code{cdata}.
If is \code{NULL} (not recommended for most cases), the user can dynamically render as many datasets views as they can source.}
}
\description{
Assemble various module components into a working one-page application for expression data
}
\details{
First, the server function calls \code{\link{multiVCtrlServer}},
which returns a named list object which either contains data or \code{NULL}.
If there is data, a \code{\link{xVServer}} is dynamically initiated to render data in its own container.
If \code{NULL}, the function removes the appropriate container using the name of the object,
which is something like "i1" and corresponds to the index of the stored global datasets.

While \code{\link{multiVCtrlServer}} is a global control that controls which datasets are displayed at all,
the \code{\link{geneVServer}} and \code{\link{selectVServer}} modules are global controls
that modify the display of any displayed datasets through applying global filters on attributes
that should be present in all or most of the data
(for expression matrix data, this means gene/protein and attributes corresponding to samples).
Thus, each \code{\link{xVServer}} component necessarily listens to
\code{\link{geneVServer}} and \code{\link{selectVServer}}, but each can also have its own
indepedent local controls, which takes precedence if enabled.
}
\seealso{
Other multiV functions: 
\code{\link{multiVUI}()}
}
\concept{multiV functions}
