% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPackR.R
\name{nest}
\alias{nest}
\title{Given a vector values in some hierarchical order (e.g. describing classes or lineages)
and data that goes under certain of these values,
this function recursively re-structures the hierarchical values into lists
and puts data into the last matching level}
\usage{
nest(v, data)
}
\arguments{
\item{v}{A vector of values to be nested, with the order corresponding to nesting levels.}

\item{leaf}{A named list containing data to be placed at the leaf level, with names matching the values in v.}
}
\value{
A nested list.
}
\description{
Given a vector values in some hierarchical order (e.g. describing classes or lineages)
and data that goes under certain of these values,
this function recursively re-structures the hierarchical values into lists
and puts data into the last matching level
}
