% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPackR.R
\name{nestD3}
\alias{nestD3}
\title{Modify nested list to make a version better corresponding to desired d3 json.}
\usage{
nestD3(context, child, contextcolors = NULL)
}
\arguments{
\item{context}{Name of the context.}

\item{child}{A list containing nested lists or data in a data.frame/data.table. See details.}

\item{contextcolors}{Optional, a named character vector mapping levels of the hierarchy to specific colors.
If not given, the context background will transparent.}
}
\value{
A list that can be converted to correctly formatted json for use with cellpack.js
}
\description{
Modify nested list to make a version better corresponding to desired d3 json.
}
