% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{networkGraph}
\alias{networkGraph}
\title{Make igraph showing connections between datasets/sources}
\usage{
networkGraph(cdir, filepattern = "*[.](txt|tsv|csv)$",
  nodes = "([^\\\\.]*)", edges)
}
\arguments{
\item{cdir}{Directory (relative to current working directory) to the collection of dataset files.}

\item{filepattern}{Pattern to help select files within the directory. Defaults to selecting all files with extension .txt|.tsv|.csv.}

\item{nodes}{Pattern for extracting node names from source file name. Defaults to using the file name without the extension.}

\item{edges}{Column within the dataset used for calculating edges, e.g. "ID".}
}
\value{
List containing igraph object and data frame summarizing nodes and edges.
}
\description{
The method first calculates node and edge counts from the collection of datasets.
Each dataset/source is a node whose size is proportional to the number of features (columns).
More concretely, when each dataset file represents a published study,
node size is proportional to the amount of data the study generated
(currently, the log of the number of features in the dataset).
Edge weights between datasets are counted as the number of overlapping IDs.
}
