% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newDataset.R
\name{newDatasetInput}
\alias{newDatasetInput}
\title{Shiny module UI for a dataset input}
\usage{
newDatasetInput(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\value{
A \code{shiny::\link[shiny]{tagList}} containing input UI.
}
\description{
An extension of \code{\link{dataUploadUI}} to include an additional text field.
}
