% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newDataset.R
\name{newDatasetServer}
\alias{newDatasetServer}
\title{Shiny module server function for handling new dataset input}
\usage{
newDatasetServer(
  id,
  datakey = NULL,
  xname = NULL,
  checkFun = NULL,
  informd = NULL,
  appdata = NULL
)
}
\arguments{
\item{datakey}{Optional, a named list containing name/label to be used in creating the key column.}

\item{xname}{Optional, a default name/label to use in the key column.}

\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\value{
A reactive data.table of the processed upload.
}
\description{
This is essentially the \code{\link{dataUpload}} module, but the uploaded data
is modified to include a key-like column before it is returned as a reactive object.
Use \code{\link{dataUpload}} if there is no need for a key,
which is used to track the dataset as in \code{\link{refSubset}}.
}
