% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridParCoordsGraph.R
\name{newHPCG}
\alias{newHPCG}
\title{New hybrid parallel coordinates graph object}
\usage{
newHPCG(data, IDcol = "ID", vcolor, vcolorX = FALSE, orderx, vlabels,
  autoscale = T, hideNA, yNAnode = -30)
}
\arguments{
\item{data}{A data.frame or data.table, in which categorical variables should be factor columns.
All character columns except \preformatted{IDcol} will be factored.}

\item{IDcol}{Name of column containing IDs, or NA if there are no ID columns. Defaults to "ID".}

\item{vcolor}{Name of the variable in data to be used for mapping to colors.}

\item{vcolorX}{Logical flag to indicate whether the variable used for \preformatted{vcolor} should be
excluded in the graph. Defaults to FALSE.}

\item{orderx}{Optional, a vector giving custom order of the axes.}

\item{vlabels}{Optional, character vector for axes labels.}

\item{autoscale}{Logical flag to indicate that numeric variables should be scaled.
Defaults to TRUE since this usually makes the best visual use of space.}

\item{hideNA}{A vector giving the factor variable names for which
\preformatted{NA} nodes should not be constructed. By default,
all factor variables with \preformatted{NA} values will display a \preformatted{NA} node.}

\item{yNAnode}{Optional, custom axis position for NA node.}
}
\description{
Create data object to be used for visualizing a dataset containing both ordinal and
categorical variables with a custom graph built using plotly, and which can be enabled for interactivity
}
\details{
Typical parallel coordinates graph work for numeric data, while another version called parallel sets
is normally used for categorical data. This "hybrid" graph is an attempt to represent both
numeric and categorical data in one graph, and as such it can be described as a typical
parallel coordinates graph constructed with floating nodes for categorical (factor) variables.
One limitation is that numeric values should be non-negative, because of how \preformatted{NA}
is handled in the plotting for the nodes. The nodes correspond to the levels within the variable,
and the sizes are proportional to the number of instances within each level.
\preformatted{NA} values in factor variables are assigned to a \preformatted{NA} node,
unless otherwise specified through \preformatted{hideNA}.
The graph is most useful when it is interactive, so it is intended to be embedded as an HTML widget.
}
