% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLink.R
\name{newOrderInput}
\alias{newOrderInput}
\title{Modified version of \code{shinyjqui::\link[shinyjqui]{orderInput}};
important changes include \code{item_class} not being constrained and allowed to be assigned to a
a custom class with a forced break.}
\usage{
newOrderInput(inputId, label, items, as_source = FALSE, connect = NULL,
  item_class, placeholder = NULL, width = "500px", ...)
}
\description{
Modified version of \code{shinyjqui::\link[shinyjqui]{orderInput}};
important changes include \code{item_class} not being constrained and allowed to be assigned to a
a custom class with a forced break.
}
