% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{nodeMod}
\alias{nodeMod}
\title{Modify network graph node aesthetics based on metadata}
\usage{
nodeMod(nodes, metadata, IDcol = "Source", mfuns, mcols)
}
\arguments{
\item{nodes}{Name of nodes to look up in metadata table.}

\item{metadata}{A \code{data.table}.}

\item{IDcol}{Column in metadata table used for matching node names. Defaults to "Source".}

\item{mfuns}{A list of one or more functions that map values in a column to size, color, or another node feature.}

\item{mcols}{A named vector of the corresponding column(s) to use as input into \code{mfuns}. Names should correspond to the node feature.}
}
\value{
A \code{data.frame} containing aesthetics data for the nodes.
}
\description{
Change or add node size, color, or other annotation values depending on metadata.
}
\details{
As asked for by the param \code{IDcol}, in metadata there should be a key column (perhaps called Source, Study, etc.)
with node IDs. Meanwhile, \code{mfuns} takes columns whose values can be mapped
to node size, color and other node properties through map functions that can be passed into \code{mfuns}.
}
