% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npodX}
\alias{npodX}
\title{A limited nPOD dataset containing variables used for matching}
\format{A data.table with 449 rows and 23 columns}
\usage{
npodX
}
\description{
This is used by \code{\link{refSubset}}
}
\keyword{datasets}
