% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLink.R
\name{orderSet}
\alias{orderSet}
\title{Create draggable and order-able set of elements}
\usage{
orderSet(vars, matchOpts, sessionNS)
}
\arguments{
\item{vars}{The subset of variables in the reference dataset used for matching.}

\item{matchOpts}{Match options, i.e. guessed pairs of matching variables for external and reference dataset.}
}
\value{
A \code{shiny::\link[shiny]{tagList}} of variables that can be dragged and dropped.
}
\description{
Create draggable and order-able set of elements
}
