% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.dive_result}
\alias{plot.dive_result}
\title{Plot an individual DIVE result}
\usage{
\method{plot}{dive_result}(x, metric, show_title = TRUE, ...)
}
\arguments{
\item{x}{A \code{dive_result}.}

\item{metric}{Alpha-diversity metric name.}

\item{show_title}{Logical.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot an individual DIVE result
}
