% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dive_heatmap}
\alias{plot_dive_heatmap}
\title{Plot all DIVE temporal associations}
\usage{
plot_dive_heatmap(result)
}
\arguments{
\item{result}{A \code{dive_result}.}
}
\value{
A ggplot object.
}
\description{
Plot all DIVE temporal associations
}
