% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dive_timeseries}
\alias{plot_dive_timeseries}
\title{Plot representative DIVE time series}
\usage{
plot_dive_timeseries(
  x,
  metrics = c("observed", "shannon", "ace", "simpson_lambda", "pielou"),
  metric_labels = c(observed = "Richness (Observed)", shannon = "Diversity (Shannon)",
    ace = "Rarity (ACE)", simpson_lambda = "Dominance (Simpson lambda)", pielou =
    "Evenness (Pielou)"),
  ncol = 2
)
}
\arguments{
\item{x}{A \code{dive_result} object.}

\item{metrics}{Character vector of alpha-diversity metrics.}

\item{metric_labels}{Optional named character vector providing
readable labels for the selected metrics.}

\item{ncol}{Number of facet columns.}
}
\value{
A ggplot object.
}
\description{
Produces separate time-series panels for selected
alpha-diversity metrics and ecosystem functionality.
}
