% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridParCoordsGraph.R
\name{plotlyHPCG}
\alias{plotlyHPCG}
\title{Construct a hybrid parallel coordinates graph}
\usage{
plotlyHPCG(hpcg, colors)
}
\arguments{
\item{hpcg}{Data from \code{\link{hybridParCoordsGraph}}.}

\item{colors}{Color mappings.}
}
\description{
Visualize a dataset containing both ordinal and categorical variables with
a custom graph built using plotly, and which can be extended with a linked pie graph
in the \code{\link{HPCGraph}} module.
}
