% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dive_result}
\alias{print.dive_result}
\title{Print a DIVE result}
\usage{
\method{print}{dive_result}(x, ...)
}
\arguments{
\item{x}{A \code{dive_result} object.}

\item{...}{Additional arguments.}
}
\value{
The object invisibly.
}
\description{
Print a DIVE result
}
