% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_md5}
\alias{raw_md5}
\title{Arrange files containing raw sequence filenames and their corresponding
MD5 hash. This files should be sorted as two column table.
MD5SUM FILENAME}
\usage{
raw_md5(extension = "md5.txt", subdirectory = "")
}
\arguments{
\item{extension}{preferred file extension, default: md5.txt}

\item{subdirectory}{subdirectory containing the files, within the current
working directory.}
}
\description{
Arrange files containing raw sequence filenames and their corresponding
MD5 hash. This files should be sorted as two column table.
MD5SUM FILENAME
}
