% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refSubset.R
\name{refSubset}
\alias{refSubset}
\title{Shiny server module function for returning a subsetted dataset}
\usage{
refSubset(input, output, session, refData, subsetfeat, refkey,
  infoRmd = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{refData}{The reference data.table.}

\item{subsetfeat}{The name of the column containing subset factors.}

\item{refkey}{Optional, a named list containing name/label for creating a key-like column,
where the name is the name of the column. See details for intended purpose.}

\item{infoRmd}{Optional, relative path to an info Rmarkdown file that can be pulled up in a modal.}
}
\value{
A reactive subsetted data.table.
}
\description{
Server function for creating a reactive subsetted dataset following user interaction.
}
\details{
When multiple instances of the module are used to create subsets from different datasets
(as opposed to subsets from the same dataset), it might be necessary to use refkey to track
the origin of the subset, and especially if the subsets are to be merged on some common
column later. For the most generic example, when datasets come from different sources,
refkey can be called "Source" and the values might be S1 and S2.
}
