% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refSubset.R
\name{refSubsetInput}
\alias{refSubsetInput}
\title{Shiny module UI for selecting a subsetted dataset}
\usage{
refSubsetInput(id, refname = NULL)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{refname}{Optional, a label for the reference dataset or even HTML link to logo, etc.}
}
\value{
A \code{shiny::\link[shiny]{tagList}} for UI to subset a dataset.
}
\description{
Shiny module UI to subset a dataset through a selection menu.
}
