% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeID}
\alias{removeID}
\title{Exclude names referring to ID in vector}
\usage{
removeID(x, ID = "ID")
}
\description{
Used with \code{names}, this filters out the assumed ID ("ID") and returns
a version used for downstream subsetting
}
\keyword{internal}
