% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{review}
\alias{review}
\title{Shiny module server for reviewing and editing table data}
\usage{
review(input, output, session, template, dt, nav = c("radioButtons",
  "selectInput"))
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{template}{UI template.}

\item{nav}{For templates containing multiple parts (sections),
this determines whether a selectInput or radioButtons type navigation menu
is generated for selecting which part of the template to be reviewed.
It is not useful for templates of length 1.}
}
\description{
The module composes the table data with the ability to edit current or new rows.
The table is connected to the collectData module,
which provides a data collection interface using a data template, the overall UI allows
editing a current or new row of data.
}
