% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{reviewUI}
\alias{reviewUI}
\title{Shiny module UI for reviewing and editing table data}
\usage{
reviewUI(id, CSS = system.file("www/", "app.css", package = "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
The UI features a basic table representation of the data and an interface for collecting data.
}
