% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVUI.R
\name{selectV}
\alias{selectV}
\title{Shiny module server for selecting columns from a data table}
\usage{
selectV(input, output, session, data, key = "ID",
  label = HTML("<strong>Phenotype/Experimental variable(s)</strong>"),
  selected = reactive({     NULL }))
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{data}{A reactive data.table.}

\item{key}{A key column that is kept for every selected subset. Defaults to "ID".}

\item{label}{Label for variable select input.}

\item{selected}{Optional, initial selection.}
}
\value{
The subsetted data table.
}
\description{
Returns the selected columns.
}
\seealso{
Other multiVApp module functions: \code{\link{geneV}},
  \code{\link{multiV}}, \code{\link{selectVUI}},
  \code{\link{subgroupVUI}}, \code{\link{subgroupV}}
}
\concept{multiVApp module functions}
