% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectV.R
\name{selectVServer}
\alias{selectVServer}
\title{Shiny module server for selecting columns from a data table}
\usage{
selectVServer(
  id,
  data = reactive({     NULL }),
  key = "ID",
  label = HTML("<strong>Phenotype/Experimental variable(s)</strong>"),
  selected = reactive({     NULL }),
  countby = reactive({     NULL }),
  maxitems = 3
)
}
\arguments{
\item{data}{A reactive data.table. See details.}

\item{key}{A key column that is kept for every selected subset. Defaults to "ID".}

\item{label}{Label for variable select input.}

\item{selected}{Optional, initial selection.}

\item{countby}{A vector of IDs, which are intersected with the key (IDs) in data
to generate counts in the select options. See details.}

\item{maxitems}{Maximum number of items that can be selected. Defaults to 3.}

\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\value{
The subsetted data table.
}
\description{
Returns the selected columns.
}
\details{
This is a subcomponent intended to serve under \code{\link{multiVCtrl}}
but should be generic enough to be integrated with other modules.
When data is uploaded through \code{\link{multiVCtrl}}, it can be "high-throughput"
or "low-throughput" type data, the latter typically as phenotype or clinical variables.
This module handles changes in the table of the latter data type.
When the user adds (uploads) new data columns, the selection updates to include these new options.
When the user selects specific columns from the menu, a column-subsetted table is returned.
The most recent modification involves adding parenthesized counts for each variable option,
e.g. "phenotypeX (15)", calculated by intersections with a vector of IDs.
}
\seealso{
Other multiVUI module functions: 
\code{\link{multiVUIR}()}
}
\concept{multiVUI module functions}
