% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectV.R
\name{selectVUI}
\alias{selectVUI}
\title{Shiny module UI for selecting columns from a data table}
\usage{
selectVUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
A basic UI for choosing columns in a dataset.
}
\seealso{
Other multiVApp module functions: 
\code{\link{subgroupVServer}()},
\code{\link{subgroupVUI}()}
}
\concept{multiVApp module functions}
