% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupVUI.R
\name{subgroupV}
\alias{subgroupV}
\title{Shiny module server for creating comparison plots between subgrouped data}
\usage{
subgroupV(input, output, session, cdata, hdlist)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{cdata}{A data.table of clinical, phenotype, or other experimental data used to define groups.}

\item{hdata}{A list of named high-dimensional (high-throughput datasets)}
}
\description{
This primarily allows creation of volcano plots for highthroughput datasets based on user-defined group contrasts.
}
\seealso{
Other multiVApp module functions: \code{\link{geneVUI}},
  \code{\link{geneV}}, \code{\link{multiVAppRun}},
  \code{\link{multiVApp}}, \code{\link{multiVUIApp}},
  \code{\link{selectVUI}}, \code{\link{selectV}},
  \code{\link{subgroupVUI}}
}
\concept{multiVApp module functions}
