% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xV.R
\name{subsetHiVar}
\alias{subsetHiVar}
\title{Subset a matrix by the most variable n features (columns)}
\usage{
subsetHiVar(data, n = NULL, percent)
}
\description{
Need to either give n number of features or percent of data to calculate n
}
\keyword{internal}
