% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dive_result}
\alias{summary.dive_result}
\title{Summarize a DIVE result}
\usage{
\method{summary}{dive_result}(object, ...)
}
\arguments{
\item{object}{A \code{dive_result} object.}

\item{...}{Additional arguments.}
}
\value{
A summary object.
}
\description{
Summarize a DIVE result
}
