% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tse}
\alias{tse}
\title{tse}
\format{
Summarized Experiment :
\describe{
\item{assay(tse)}{Abundance data}
\item{rowData(tse)}{heirarchal ranking}
\item{colData(tse)}{samples data}
}
}
\source{
Oosterhorn sampling campaign.
}
\usage{
tse
}
\description{
Abundance data of features in community (32 samples).
}
\keyword{datasets}
