% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{validateRefData}
\alias{validateRefData}
\title{Validate refdata object used in matchApp}
\usage{
validateRefData(dataset, customdata = "Cohort", subsetv = NULL)
}
\arguments{
\item{dataset}{A `data.table`.}

\item{customdata}{Attribute column name, defaults to "Cohort".
This should be changed if matching on something else, e.g. "Sites".}

\item{subsetv}{Optional, the name of the column storing subset attribute
for subsetted matching.}
}
\description{
The refdata object is required in \code{\link{matchAppServer}}.
For now, this calls an internal function to check that whatever is used
as refdata contains numeric/factor-encoded attributes only,
except for the fields given by customdata and subsetv.
}
