% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBox.R
\name{valueBox}
\alias{valueBox}
\title{A UI box element for summary values in dashboard-like presentation}
\usage{
valueBox(
  value,
  subtitle,
  icon = NULL,
  textcolor = "black",
  bgcolor = "white",
  width = 4,
  href = NULL
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
This is similar to the UI element in shinydashboard.
}
