% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{vBinColor}
\alias{vBinColor}
\title{Map a binary vector to a color. For example, the vector \preformatted{c(T, F, F)}
could be assigned the color white if any value is true (true is dominant),
or black if any value is false (black is dominant),
or one could use the proportion of true/false to get some gray color.}
\usage{
vBinColor(x, domlogic = T, shades = c("black", "white"))
}
\arguments{
\item{x}{A logical vector.}

\item{domlogic}{Dominant logical. Defauls to true.}

\item{shades}{A vector of colors with at least two colors, where the last color is mapped to the dominant logical.
Defaults to "black" and "white". If more than two colors are given, the proportion of the dominant logical in x is used
to assign the closest shade.}
}
\description{
Map a binary vector to a color. For example, the vector \preformatted{c(T, F, F)}
could be assigned the color white if any value is true (true is dominant),
or black if any value is false (black is dominant),
or one could use the proportion of true/false to get some gray color.
}
