% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{visNetworkMake}
\alias{visNetworkMake}
\title{Wrapper to update network graph}
\usage{
visNetworkMake(network, aes)
}
\arguments{
\item{network}{An igraph object.}

\item{aes}{Data frame containing nodes for which aesthetics are added or modified.}
}
\value{
A visNetwork object.
}
\description{
Apply specified aesthetics to igraph object and exports as
interactive visNetwork object for embedding in HTML/Shiny application.
}
