% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{vSize}
\alias{vSize}
\title{To get size, a numeric column can be summed by each node, which is then usually log-scaled.
Note: For simplicity, NA is substituted by 1. Because sizes are log-scaled and only meant to proportionally representative,
this doesn't matter too much.}
\usage{
vSize(x, logscale = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{scale}{Whether to log-scale node sizes for aesthetic reasons.}
}
\description{
To get size, a numeric column can be summed by each node, which is then usually log-scaled.
Note: For simplicity, NA is substituted by 1. Because sizes are log-scaled and only meant to proportionally representative,
this doesn't matter too much.
}
