% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{xCheckID}
\alias{xCheckID}
\title{xCheckID
Also used for match module
Returns a list of IDs found in reference list
#' Tells user that internal matches have been found}
\usage{
xCheckID(data, ref = cdata$ID)
}
\description{
xCheckID
Also used for match module
Returns a list of IDs found in reference list
#' Tells user that internal matches have been found
}
