% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xV_classes.R
\name{xVExprs}
\alias{xVExprs}
\title{Helper to construct dataset object for use in \code{\link{xVServer}} module}
\usage{
xVExprs(x, title = NULL, altj = NULL, type = "transcriptomics")
}
\arguments{
\item{x}{A numeric matrix that should have F features as columns and S samples as rows,
with appropriate names/IDs as column names and row names.}

\item{title}{Character title for the dataset.}

\item{altj}{A character or integer vector matching length of features (columns),
intended to be used as an interoperable alternate index.}

\item{type}{Either "genomics" or "proteomics", currently.}
}
\description{
Helper to construct dataset object for use in \code{\link{xVServer}} module
}
