% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xVUI.R
\name{xVUI}
\alias{xVUI}
\title{Shiny module UI for presenting high dimensional data with other features in multi-column view}
\usage{
xVUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
UI based on the idea of a "visual spreadsheet" or "line up" view. The associated
server function \code{\link{multiV}} provides more details of the implementation.
}
\details{
The original use case presumes the high dimensional data to be genomics data, e.g.
gene or protein expression that can be represented with heatmap matrices.
}
