% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browseUI}
\alias{browseUI}
\title{Shiny app UI for browsing metadata}
\usage{
browseUI(id, CSS = system.file("www/", "app.css", package = "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}
}
\description{
Shiny app UI for browsing metadata
}
