% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFromCollection.R
\name{cdataMake}
\alias{cdataMake}
\title{Make data object for Shiny app}
\usage{
cdataMake(cdir = "./Collection/Main/", filepattern = "*.txt",
  exclude = "!", other = list("./Collection/Other/coredata_ref.txt"),
  outdir = "./App/Data/")
}
\arguments{
\item{cdir}{Directory path (relative to working directory) to the collection of dataset files.}

\item{filepattern}{Pattern for grep to identify qualifying files in the directory, e.g. "*.txt".}

\item{exclude}{A pattern for grep to exclude columns within a file from the final master dataset.}

\item{other}{A list of of paths to other data files not in the main collection but
that should be part of the master data object; an empty list or "" if no other data to be incorporated.}

\item{outdir}{Where to put \preformatted{cdata}. Defaults to App/Data.}
}
\description{
\code{cdataMake} is a wrapper to create the \code{cdata} object
}
