% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPack.R
\name{cellPack}
\alias{cellPack}
\title{Shiny server module function for cellPack widget}
\usage{
cellPack(
  input,
  output,
  session,
  json,
  css = system.file("www/", "cellpack.css", package = "DIVE")
)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{json}{Either a JSON object or relative path to the JSON data file for widget.}

\item{css}{Optional, relative path to a CSS file.}
}
\value{
Reactive object for element that was clicked
}
\description{
Shiny server module function for cellPack widget
}
