% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPack.R
\name{cellPackUI}
\alias{cellPackUI}
\title{Shiny module UI widget for visualization and filtering of hierarchical cell and tissue samples}
\usage{
cellPackUI(id, width = "400px", height = "400px")
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Based on d3 circlepack.js
}
\details{
Certain selections can trigger the widget "augmented input interface".
}
