% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkCohortData}
\alias{checkCohortData}
\title{Check match data
Checks that data has an ID column and that data is numeric/numeric factors}
\usage{
checkCohortData(cohdata, message = NULL, result = NULL)
}
\description{
Check match data
Checks that data has an ID column and that data is numeric/numeric factors
}
