% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLink.R
\name{consortSet}
\alias{consortSet}
\title{Create draggable and order-able set of elements}
\usage{
consortSet(varset, matchOpts, sessionNS)
}
\arguments{
\item{matchOpts}{Optional list of guessed pairs of harmonized variables for external and reference dataset.}

\item{sessionNS}{Namespace function, i.e. \code{shiny::\link[shiny]{session$ns}}, important for a namespaced context.}

\item{vars}{The subset of variables in the reference dataset used for matching.}
}
\value{
A \code{shiny::\link[shiny]{tagList}} ui for drag-n-drop variables.
}
\description{
Create draggable and order-able set of elements
}
