% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastV.R
\name{contrastVUI}
\alias{contrastVUI}
\title{Shiny module contrast UI}
\usage{
contrastVUI(id, choices)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{choices}{A vector corresponding to choices of grouping variables for contrast.}
}
\description{
UI for user to define contrast groups for comparison plots
}
\seealso{
Other contrastV module functions: 
\code{\link{contrastVServer}()}
}
\concept{contrastV module functions}
