% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2cor.R
\name{cordataMake}
\alias{cordataMake}
\title{Wrapper to instantiate correlation data for Shiny app}
\usage{
cordataMake(cdata)
}
\arguments{
\item{cdata}{A data.table}
}
\value{
A list with 1) corM, the correlation matrix, and 2) corN, the number of samples available
when r was calculated using pairwise complete observation
}
\description{
Wrapper to instantiate correlation data for Shiny app
}
