% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2cor.R
\name{data2cor}
\alias{data2cor}
\title{Get a correlation matrix and shared samples for each correlation}
\usage{
data2cor(cdata, exclude = "^ID|_SE$|_SEM$|_SD$")
}
\arguments{
\item{cdata}{A data.table}

\item{exclude}{Optional, pattern of columns to exclude, e.g. columns containing IDs, irrelevant or non-numeric values.}
}
\value{
A list with M, the correlation matrix, and N, the number of samples.
}
\description{
After filtering the input data, get Spearman r using pairwise complete observations
as well as the number of pairwise observations available.
}
