% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{dataHelperServer}
\alias{dataHelperServer}
\title{Shiny module server function for simple interactions between two relational tables}
\usage{
dataHelperServer(
  id,
  lhdata,
  lhdatakey = "ID",
  rhdata,
  rhdatakey = "ID",
  initcolselect = NULL,
  handler,
  filterconf
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{lhdata}{A \code{data.table} of the dataset on the left-hand side.}

\item{lhdatakey}{Name of unique key column in \code{lhdata}, defaulting to "ID".}

\item{rhdata}{A \code{data.table} of the dataset on the right-hand side.}

\item{rhdatakey}{Name of unique key column in \code{rhdata}, defaulting to "Source".}

\item{initcolselect}{Which columns in \code{rhdata} should be shown initially,
as determined by a vector of names matching cols in \code{rhdata}.
A checkboxgroup element for column selection will also be rendered.
By default \code{NULL} so there is no selection control and all columns will be shown.}

\item{handler}{A \code{data.table} used for translating keys between \code{lhdata} and \code{rhdata}.}

\item{filterconf}{If given, will create filters for the left-hand dataset using \code{\link{sideFilterUI}}.
Should be a list with names matching columns in \code{lhdata} to be used as filters
and elements "type" with their input widget types, "selected" for the initial selection,
and "conditional" with a boolean value for whether this is a filter that should initially be hidden/ignored.}
}
\description{
Implement interface to traverse data in "righthand" and "lefthand" tables with additional attribute filters
}
\details{
The simple interface implemented allows general browsing, filtering, subsetting of traditional tabular data,
limited to two tables. The module requires these tables:
\describe{
  \item{\code{lhdata}}{ "Left-hand" table of data.
  Originally, this was a table of case IDs with columns for attributes such as age and sex. }
  \item{\code{rhdata}}{ "Right-hand" table of data.
  Originally, this was a table of studies with columns for attributes such as authors, description and reference.}
  \item{\code{handler}}{ Lookup table to translate relations between the left-hand and right-hand tables. }
}
}
\seealso{
Other dataHelper: 
\code{\link{dataHelperUI}()}
}
\concept{dataHelper}
