% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{dataHelperServer}
\alias{dataHelperServer}
\title{Shiny module server for simple traversal of relational tables in a database}
\usage{
dataHelperServer(
  id,
  dbcon = NULL,
  lhdata,
  lhdatakey,
  rhdata,
  rhdatakey,
  handler
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{dbcon}{A database connection object to the database containing the required tables.}

\item{lhdata}{Name of the table on the left-hand side.}

\item{lhdatakey}{Name of key column in \code{lhdata} table, which should also be a key in \code{handler}.}

\item{rhdata}{Name of the table on the right-hand side.}

\item{rhdatakey}{Name of key column in \code{rhdata} table, which should also be a key in \code{handler}.}

\item{handler}{Name of table used for translating between \code{lhdata} and \code{rhdata}.}
}
\description{
Traverse data in "right-hand" and "left-hand" tables
}
\details{
Implements simple interface allowing constrained browsing, filtering, subsetting of traditional tabular data.
The module requires these tables:
\describe{
  \item{\code{lhdata}}{ "Left-hand" table of data. Conceptually, also known as a "dimension table".
  For example, a table of unique patients and attributes such as age, sex, race, etc. }
  \item{\code{rhdata}}{ "Right-hand" table of data. Conceptually, also known as a "dimension table".
  For example, a table of unique diagnoses and attributes such as description, insurance code, etc. }
  \item{\code{handler}}{ Lookup table to translate relations between the left-hand and right-hand tables.
  Using current example, a table linking patient IDs to diagnosis IDs (which can be many-to-many).
  See helpful reference for
  \href{https://www.ibm.com/support/knowledgecenter/SSEP7J_11.1.0/com.ibm.swg.ba.cognos.ug_fm.doc/c_dyn_query_bridge_tables.html}{bridge table}}.
}
}
\seealso{
Other dataHelper: 
\code{\link{dataHelperUI}()}
}
\concept{dataHelper}
