% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper.R
\name{dataHelperUI}
\alias{dataHelperUI}
\title{Shiny app UI for simple traversal of relational tables in a database}
\usage{
dataHelperUI(
  id,
  CSS = system.file("www/", "app.css", package = "DIVE"),
  theme = "paper",
  oneway = FALSE,
  lefticon = NULL,
  righticon = NULL
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}

\item{theme}{Optional, name of theme for \code{shinythemes}.}

\item{oneway}{Whether data traversal should be one-way (left-to-right) only, or both ways.
Defaults to \code{FALSE}.}

\item{lefticon}{Optional, an icon to represent the left-hand data.}

\item{righticon}{Optional, an icon to represent the right-hand data.}
}
\description{
UI to traverse data between "right-hand" and "left-hand" tables
}
\seealso{
Other dataHelper: 
\code{\link{dataHelperServer}()}
}
\concept{dataHelper}
