% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUpload.R
\name{dataUpload}
\alias{dataUpload}
\title{Shiny server function for data upload module}
\usage{
dataUpload(input, output, session, removable = F, infoRmd = NULL,
  appdata = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{removable}{Logical flag to indicate whether data upload will have "removable" feature. Defaults to FALSE. See details.}

\item{infoRmd}{Optional, an Rmarkdown help file for infoOutput, e.g. requirements.}

\item{appdata}{Optional, the name (including extension) of one or more files stored in appdata that can be
mock-uploaded. See details.}
}
\value{
The uploaded file as a reactive data.table object.
}
\description{
The most basic implementation is to process and return any input to fileInput in the UI.
File uploads can have "reset" behavior if the removable optional feature is specified,
where a remove button will be rendered after upload.
The module also optionally incorporates \code{\link{infoOutput}} functionality.
Finally, it is possible to perform a mock upload of a saved dataset, e.g. for demonstration purposes.
Saved datasets are expected to be .csv files that reside in a relative "Data/" directory.
For instance, if the name is "SampleCohort", the dataset will be "uploaded" from "Data/SampleCohort.csv".
}
