% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUpload.R
\name{dataUploadUI}
\alias{dataUploadUI}
\title{Shiny module UI for user-uploaded data}
\usage{
dataUploadUI(id, ...)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{...}{Named list of params passed to \code{shiny::\link[shiny]{fileInput}}.
Note that \code{multiple = FALSE} (the default) must not be overridden.}
}
\value{
UI components.
}
\description{
Create UI that contains a main \code{shiny::\link[shiny]{fileInput}}
and two optional features: a file-remove button and link to display more info
(for communicating file upload requirements and/or instructions).
}
\seealso{
Other dataUpload functions: 
\code{\link{dataUploadServer}()}
}
\concept{dataUpload functions}
