% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meh.R
\name{dev_mode}
\alias{dev_mode}
\title{Internal function for getting and setting dev mode}
\usage{
dev_mode(setting = NULL)
}
\arguments{
\item{setting}{TRUE or FALSE, passed to \code{base::options} \code{DIVE.dev.mode}.}
}
\description{
Control the display of error messages in \code{\link{meh}} and of certain other UI
}
\details{
Aside from \code{\link{meh}}, some modules use this setting:
\itemize{
  \item \code{\link{matchResultServer}}: intermediate results are only downloadable when ON
}
}
