% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive.R
\name{dive}
\alias{dive}
\title{Diversity-Informed Valuation of Ecosystem functionality}
\usage{
dive(tse, functionality, output_dir = "figures")
}
\arguments{
\item{tse}{abundance data of features in community (format: summarized experiment with n sample)}

\item{functionality}{ecosystem performance measure (format: dataframe dimensions n*1)}

\item{output_dir}{directory to print cross correlations function in}
}
\value{
dataframe containing varying alpha diversity metrics
}
\description{
Diversity-Informed Valuation of Ecosystem functionality
}
\examples{
dive(tse, functionality, output_dir = "figures")
}
