% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIVE.R
\docType{package}
\name{DIVE}
\alias{DIVE}
\alias{{pkgname}-package}
\title{DIVE: Data Integration and Visual Exploration}
\description{
The DIVE package is an approach to integration and interactive visual exploration of linked datasets
and implements Shiny modules that can be run as stand-alone apps or part of a larger application.
}
\seealso{
\href{https://github.com/avucoh/DIVE}{Package GitHub}
\subsection{Modules}{
\subsection{Match module}{

\code{\link{matchApp}}
}

\subsection{Matrix Module}{

\code{\link{matrixApp}}
}

\subsection{MultiV Module}{

\code{\link{multiVUIR}}
}

\subsection{Data Helper Module}{

\code{\link{dataHelperServer}}
}

}

\subsection{Package options}{

(TO-DO)
}
}
