% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIVE.R
\docType{package}
\name{DIVE}
\alias{DIVE}
\title{DIVE: Data Integration and Visual Exploration}
\description{
The DIVE package implements an approach to integration and interactive visual exploration of biological datasets with shared samples. It includes five Shiny modules that can be run as stand-alone apps or part of a larger application. \url{https://github.com/avucoh/DIVE}
}
\section{Matching Module}{

\code{\link{matchAppR}}
}

\section{Matrix Mix Module}{

\code{\link{matrixMixR}}
}

\section{MultiVUI Module}{

\code{\link{multiVUIR}}
}

\section{Browse Module}{

\code{\link{browseR}}
}

