% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excludePresent}
\alias{excludePresent}
\title{Exclude values in dataset #2 that are present in dataset #1}
\usage{
excludePresent(x = "ID", data1 = NULL, data2 = NULL)
}
\description{
Return a version of dataset #2 that doesn't contain records in dataset #1,
assuming records are in an "ID" column
}
\keyword{internal}
