% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelper2.R
\name{filterPanel2Server}
\alias{filterPanel2Server}
\title{Filter server}
\usage{
filterPanel2Server(id, dt, dtkey = "ID", filterconf)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{dt}{A database table.}

\item{dtkey}{Name of unique key column in \code{dt}, defaulting to "ID".}

\item{filterconf}{A list with names of columns in \code{dt} to be used as filters,
elements "input" with their input widget types, "selected" for the initial selection,
and "conditional" with a boolean value for whether
this is a filter that should initially be active/inactive.}
}
\description{
Render filter ui and process output from user interaction
}
\details{
The server module dynamically renders a filter ui given the specification
passed in using \code{filterconf}. Because not all available filters need to be "active",
\code{filterconf} allows specifying which filters should be initially active
and which should be considered optional. Users can toggle which filters to apply.
Each input component filters upon values of columns in the table \code{dt}.
The applied combination is outputted as matching \code{dtkey} entries
(not entire rows). Calculation is actually delegated to the function \code{filter4j}.
}
