% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneV.R
\name{geneVUI}
\alias{geneVUI}
\title{Shiny module UI for select-filter of genes or gene products}
\usage{
geneVUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Augmented select-filter interface for genes
}
\seealso{
Other geneV functions: 
\code{\link{geneVServer}()}
}
\concept{geneV functions}
