% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneVUI.R
\name{geneVUI}
\alias{geneVUI}
\title{Shiny module UI for filtering high dimensional genomics data}
\usage{
geneVUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Currently, this provides options to subset genomics data using a custom list input
or GO and Reactome annotations.
}
\seealso{
Other multiVApp module functions: \code{\link{geneV}},
  \code{\link{multiVAppRun}}, \code{\link{multiVApp}},
  \code{\link{multiVUIApp}}, \code{\link{selectVUI}},
  \code{\link{selectV}}, \code{\link{subgroupVUI}},
  \code{\link{subgroupV}}
}
\concept{multiVApp module functions}
