% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO.R
\name{getGEOInput}
\alias{getGEOInput}
\title{Shiny module UI for importing data from GEO}
\usage{
getGEOInput(id, informd = system.file("help/GEO_module.Rmd", package = "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{informd}{Optional path to Rmarkdown file containing help-text or similar to displayed.}
}
\description{
Interactively import GEO data with step-by-step process using a series of modals
}
