% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEO.R
\name{getGEOServer}
\alias{getGEOServer}
\title{Shiny module server for importing data from GEO}
\usage{
getGEOServer(id)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\value{
GEOdata as reactive values object. See details.
}
\description{
Handles importing GEO data by guiding user through a number of modals.
}
\details{
Returns a reactive values object containing \preformatted{$accession},
\preformatted{$eset}, \preformatted{$pData} and \preformatted{$call}.
The variable \preformatted{$call} is an internal counter for use with
other modules that need to update when new GEO dataset has been processed
completely through the annotation step
(even when it has the same accession number).
}
