% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLink.R
\name{getLinkInput}
\alias{getLinkInput}
\title{Shiny module UI for retrieving internal data with action buttons/links}
\usage{
getLinkInput(id, labels, link = T)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{labels}{One or more labels.}

\item{link}{Logical flag to indicate whether element is action link or button. Defaults to link.}
}
\value{
A div were UI is inserted.
}
\description{
Puts any number of action buttons/links on page that can be used to 
get data from a specified source.
}
