% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{hdlistMake}
\alias{hdlistMake}
\title{Wrapper to create hdlist (data object for multiV module)}
\usage{
hdlistMake(
  datadir = "data/tsv/",
  annodir = "ext-annotation/GeneID/",
  indexfile = "build/xV_index.tsv"
)
}
\arguments{
\item{datadir}{Directory where datasets live (with ending slash).}

\item{annodir}{Directory where extended annotations live (with ending slash).}

\item{indexfile}{File used to select data for conversion;
must have fields "Title" "DatasetID" "Ref" "Type"; other metadata ignored.}
}
\description{
Transforms data and annotation files to \code{xVExprs} class objects
which are passed in as `hdlist` to \code{\link{multiVServer}} module,
}
\details{
This requires an index file, e.g. xV_index.tsv, which indexes files and attributes
that need conversion to \code{\link{multiVServer}} application data, then calls \code{DIVE::xVExprs} to do so.
}
