% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{hdlistMake}
\alias{hdlistMake}
\title{Wrapper to create hdlist (data object for multiV module)}
\usage{
hdlistMake(indexfile = "index.yml")
}
\arguments{
\item{indexfile}{Path to yaml configuration file used for data conversion.
Must have fields "Title", "Ref", "Type", "DataPath", and "AnnotationPath",
where "DataPath" contains path to the data and "AnnotationPath" contains
path to the annotation file; other metadata are ignored.}
}
\description{
Transforms data and annotation files to \code{xVExprs} class objects
which are passed in as `hdlist` to \code{\link{multiVServer}} module,
}
\details{
This reads an index yaml configuration file that
stores files and attributes for data that
should be converted to \code{\link{multiVServer}} application data;
\code{DIVE::xVExprs} is called to do the conversion.
}
