% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoOutput.R
\name{infoOutput}
\alias{infoOutput}
\title{Shiny module UI to display a modal with additional info}
\usage{
infoOutput(id, label = "data req's", i = "exclamation-circle")
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{label}{Label, typically for specifying type of info.}

\item{i}{Icon type to display next to label.}
}
\value{
An actionLink.
}
\description{
Shiny module UI to display a modal with additional info
}
