% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{infoServer}
\alias{infoServer}
\title{Server module function to display a modal on click}
\usage{
infoServer(id, informd)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{informd}{Relative path to the Rmarkdown file, whose contents will be displayed in the modal.}
}
\description{
Display content from Rmarkdown file when link or button is clicked
}
