% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveMatrix.R
\name{interactiveMatrixUI}
\alias{interactiveMatrixUI}
\title{Shiny module UI of matrix plot with drilldown components}
\usage{
interactiveMatrixUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Creates app UI for a heatmap matrix responding to inputs, new plot data, that is linked to a drilldown scatterplot component.
}
