% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFromCollection.R
\name{makeCdata}
\alias{makeCdata}
\title{Make data object for Shiny app}
\usage{
makeCdata(
  cdir = "./Collection/Main/",
  filepattern = "*.txt|.csv|.tsv",
  other = NULL
)
}
\arguments{
\item{cdir}{Directory path (relative to working directory) to the collection of dataset files.}

\item{filepattern}{Pattern for grep to identify qualifying files in the directory, e.g. "*.txt".}

\item{other}{Optional, a list of of paths to other data files not in the main collection but
that should be part of the master data object. NULL if no other data to be incorporated.}
}
\description{
A wrapper to create the \code{cdata} object
}
