% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFromCollection.R
\name{makeFromCollection}
\alias{makeFromCollection}
\title{Create master data table from a collection of datasets}
\usage{
makeFromCollection(cdir, filepattern, exclude, index)
}
\arguments{
\item{cdir}{Directory path (relative to working directory) to the collection of dataset files.}

\item{filepattern}{Pattern for grep to identify qualifying files in the directory, e.g. "*.txt".}

\item{exclude}{A pattern for grep to exclude columns within a file from the final master dataset.}

\item{index}{Optional function for generating a custom index format from file names,
especially useful if naming scheme of the collection files changes.}
}
\value{
A "master" data.table
}
\description{
Builds one large master dataset given the directory where a collection of datasets resides.
This data object is required for the Shiny application.
}
