% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchApp}
\alias{matchApp}
\title{Shiny app server for matching and exploration of two datasets}
\usage{
matchApp(input, output, session, refdata, HPCG, colors, datakey, xname,
  refname, vars, guess = DIVE::guessMatch, subsets, subsetfeat,
  informd = system.file("help/cohort_exchange.Rmd", package = "DIVE"),
  factorx, appdata = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{refdata}{The reference dataset.}

\item{HPCG}{Optional graph object representing the reference dataset.}

\item{colors}{Colors}

\item{datakey}{Passed to \code{refSubsetInput}, \code{newDataset}, \code{matchResult}.}

\item{vars}{Optional, variables allowed to be used for matching within \code{refdata}, organized in categories
displayed as separate lists (mainly for user-friendliness). If not given, all variables can be used and will be displayed in one list.}

\item{guess}{Optional, a function for making initial guesses of matchable variables.}

\item{subsetfeat}{Which variable to be used as the subset variable.}

\item{informd}{Link to a help file that can be called with the info module.}

\item{appdata}{See ?}
}
\description{
Shiny app server for matching and exploration of two datasets
}
\details{
This server function puts together a number of modular server module components with the correct
logic to power the interactive capabilities of the matching application.
}
