% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchApp}
\alias{matchApp}
\title{Launch Shiny app for matching between two datasets}
\usage{
matchApp(ns = NULL, ...)
}
\arguments{
\item{ns}{Optional namespace for app, passed into server module.}

\item{...}{Parameters passed into server modules.}
}
\description{
Wrapper to launch app at console
}
