% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchAppServer}
\alias{matchAppServer}
\title{Shiny app server for matching and exploration of two datasets}
\usage{
matchAppServer(
  id,
  refname = "Reference Cohort",
  refdata,
  subsetv,
  subsets,
  customdata = "Cohort",
  defaultvalue = "CohortX",
  vars,
  guess = NULL,
  informd = system.file("info/cohort_exchange.Rmd", package = "DIVE"),
  appdata = NULL
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{refname}{Name of the reference dataset, used for labels. Defaults to "Reference Cohort".}

\item{refdata}{Reference dataset that can be matched against in whole or in part, i.e. the match pool.}

\item{subsetv}{Name of the subset-by variable in the data.}

\item{subsets}{Optional, a list to display groups with higher-level grouping in the selection.
By default, the selection is composed of unique values in \code{subsetv}.}

\item{customdata}{A name for the appended colum to user-uploaded input data.
This depends on the type of data expected. "Cohort" is the provided default for cohort-matching application.
See \code{\link{customDatasetServer}}.}

\item{defaultvalue}{Default attribute value to append to user-uploaded input data.
"CohortX" is the provided default for cohort-matching application. See \code{\link{customDatasetServer}}.}

\item{vars}{A named list of a variable set (or sets) in the data that is allowed for matching. See details.}

\item{guess}{Optional, name of the function to call for initial guessing of harmonized variables. See details.}

\item{informd}{Relative path to the Rmarkdown file, whose contents will be displayed in the modal.}

\item{appdata}{Optional, the path of one or more files that can be mock-uploaded. See details.}
}
\description{
Shiny app server for matching and exploration of two datasets
}
\details{
This server function puts together modular components to
implement the whole interaction of the matching application with several defaults
to facilitate quick startup.
Though originally conceived for matching and exploration of patient cohort datasets,
this could be adapted to matchable data in other domains, such as geographic sampling sites, etc.
This app module can also be viewed as a starting template to help
compose a new matching app with a sufficiently different layout or functionality.
}
