% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchAppUI}
\alias{matchAppUI}
\title{Shiny app UI for matching application}
\usage{
matchAppUI(
  id,
  CSS = system.file("www/", "app.css", package = "DIVE"),
  refname = NULL,
  addonUI = NULL,
  addontab = ""
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to set look and feel of the module.}

\item{addonUI}{Optional, ui module function to display addon components.}

\item{addontab}{If addonUI is given, should also give a name for the tab hosting addon ui components.}
}
\description{
Shiny app UI for matching application
}
\details{
The application UI can be viewed as a default template of how to put together several module components,
including \code{\link{newDataSetInput}}, \code{\link{refSubsetInput}}, \code{\link{cohortGraphOutput}},
to overall enable interactive exploration and matching of two different datasets. Note that though the datasets
are typically conceived as datasets of matchable patient cohorts, this could be other types of matchable data
such as geographic sampling sites, etc.
}
