% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchApp.R
\name{matchAppUI}
\alias{matchAppUI}
\title{Shiny app UI for matching application}
\usage{
matchAppUI(
  id,
  CSS = system.file("www/", "app.css", package = "DIVE"),
  refname = NULL,
  addonUI = NULL,
  addontab = ""
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to set look and feel of the module.}

\item{refname}{Optional, name of the reference dataset.}

\item{addonUI}{Optional, ui module function to display addon components.}

\item{addontab}{If addonUI is given, should also give a name for the tab hosting addon ui components.}
}
\description{
Shiny app UI for matching application
}
\details{
The application UI puts together multiple module components to enable
exploration and matching of two different but same-type datasets.
It places a \code{customDatasetInput} component for a user-uploaded input dataset in the top left
and a \code{dataSubsetInput} component to allow selection of the match pool in the top right.
The bottom half is a tab panel with tabs dynamically shown for a matching interface,
result table and result plots.
}
