% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLink.R
\name{matchLinkServer}
\alias{matchLinkServer}
\title{Server module function for interactive drag-and-drop variable harmonization of two datasets}
\usage{
matchLinkServer(
  id,
  refdata,
  setX,
  vars,
  guess = NULL,
  informd = system.file("help/matching_methods.Rmd", package = "DIVE")
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{refdata}{Reactive subsetted reference data.table.}

\item{setX}{Reactive data.table dataset, which typically comes from the \code{\link{newCohortInput}} module.}

\item{vars}{A named list of a variable set (or sets) in the data that is allowed for matching. See details.}

\item{guess}{Optional, name of the function to call for initial guessing of harmonized variables. See details.}

\item{informd}{Optional, relative path to an info Rmarkdown file that will be pulled up in a modal.
Recommend using for providing methods details of the matching algorithm.}
}
\value{
Reactive list of parameter values with \code{run}, \code{matchOpts} and \code{matchOn}. See details.
}
\description{
Given two datasets, a "reference" and a "comparison" dataset, the module implements an interface
that allows users to interactively harmonize variables to be used for matching;
the output is used as inputs to the module \code{\link{matchResult}}:
\code{params$matchOpts} contains all the possible variables that could be used between the two datasets,
\code{params$matchOn} contains variables that are actually chosen for matching, and
\code{params$run} reactively transmits the "finalized" params/triggers the \code{\link{matchResult}} module.
The drag-and-drop interface allows the user to harmonize variables manually, e.g. the "gender" label
can be dropped next to the "sex" label to say that these mean the same thing. Which variables
the user can interact with is determined by the vars parameter.
}
\section{Variables List}{

This should be a list such as \preformatted{list(Variables = c("age", "sex", "BMI", "biomarkerLevel"))} or
\preformatted{list(Demographic = c("age", "sex"), Clinical = c("BMI", "biomarkerLevel"))}.
This format allows visually grouping the variables in the display only and has no effect on matching.
}

\section{Guess Function}{

A guess function can be provided to attempt guessing which variables are actually the same between the datasets,
which will run before the user does manual harmonization of variables. The guess function is specific
to the dataset and the domain. For example, it knows that "sex" is a matchable variable in the reference dataset
and can try to look for variables called "sex" or "gender" in the incoming dataset so that the user doesn't
have to specify this manually. It should also be consistent with what is specified as matchable variables,
so that its return (a list) should contain a list item for all vars.
}

