% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchResult.R
\name{matchPair}
\alias{matchPair}
\title{Matching main function}
\usage{
matchPair(data, groupcol, matchon)
}
\arguments{
\item{data}{The data.}

\item{groupcol}{Name of the column containing groups to match between.}

\item{matchon}{Features to match on.}
}
\value{
A list containing result and a table where each row corresponds to a matched pair.
}
\description{
Calls \code{\link[optmatch]{pairmatch}} to perform matching using the desired parameters.
}
