% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPlot.R
\name{matchPlotServer}
\alias{matchPlotServer}
\title{Shiny module server for comparison histodot plots of attributes in two datasets}
\usage{
matchPlotServer(id, s1data, s2data, results)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{s1data}{Reactive data from source #1 (often the reference dataset).}

\item{s2data}{Reactive data from source #2, such as from \code{\link{customDatasetServer}}.}

\item{results}{Optional, reactive return value of \code{\link{matchResultServer}}.}
}
\description{
Histodot plots of two datasets side-by-side, with or without match information
}
\details{
The module was implemented to compare two cohort datasets but can also allow comparison
of other types of datasets that are conceptually similiar.
For the original cohort matching use case, if there happens to be match return values
from \code{\link{matchResultServer}} passed to \code{results},
the plots will add color-coding to visualize which individuals are matched.
}
\seealso{
Other matchPlot functions: 
\code{\link{matchPlotUI}()}
}
\concept{matchPlot functions}
