% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchResult.R
\name{matchResultOutput}
\alias{matchResultOutput}
\title{Shiny module UI output for match results}
\usage{
matchResultOutput(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\value{
Results UI consisting of a table and data download buttons.
}
\description{
Shiny module UI output for match results
}
