% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchResult.R
\name{matchResultServer}
\alias{matchResultServer}
\title{Shiny module server for generating match results output}
\usage{
matchResultServer(id, refdata, inputdata, params, sourcecol)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{refdata}{Reactive refdata dataset as a \code{data.table}.}

\item{inputdata}{Reactive \code{data.table} of the comparison dataset, e.g from \code{\link{customDatasetServer}}.}

\item{params}{Reactive parameters data, e.g. from \code{\link{matchLinkServer}}.}

\item{sourcecol}{Name for the source key column of the joined dataset.}
}
\value{
Reactive values containing params, intermediate results, pair, and matchtable.
}
\description{
Carries out matching and returns a match result reactive object
}
\details{
The module initializes a \code{results} reactive values list containing \code{NULL} values and
whenever new \code{inputdata} is passed in (clearing previous results).
The matching doesn't run until a valid \code{params} reactive object is provided,
which will call upon the underlying handler \code{matchPair}.
If successful, the module updates and returns the \code{results} reactive values list object
with \code{results$params} \code{results$intermediate} \code{results$pair} \code{results$matchtable}.
Though \code{results$matchtable} is rendered in this module's UI with its own download handler,
it and the rest are relevant data necessarily forwarded to \code{\link{matchPlotServer}}.
}
