% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixApp.R
\name{matrixApp}
\alias{matrixApp}
\title{Launch Shiny app for exploration of relationships in annotated data with an interactive matrix}
\usage{
matrixApp(ns, ...)
}
\description{
Wrapper to launch app at console
}
