% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCtrl.R
\name{matrixCtrl}
\alias{matrixCtrl}
\title{Shiny module server functions to generate filter UI for interactive matrix}
\usage{
matrixCtrl(input, output, session, M, N, cdata, metadata,
  newdata = reactive({ }), widget = reactive({ }))
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{M}{A non-reactive data matrix, e.g. a correlation matrix, which must have variables as rownames.}

\item{N}{A non-reactive matrix of the same dimensions as M with data for the filterable layer, e.g. sample size.}

\item{cdata}{The non-reactive data used for generating the matrix.}

\item{metadata}{A data.table with "Variable" as a key column and any number of columns (metadata) to be used as filters.}

\item{newdata}{Optional, reactive data, e.g. from user upload, that can be merged with M.}

\item{widget}{Optional, associated widget input. See details.}
}
\value{
Reactive state values to be used by the associated plotting module.
}
\description{
Update/populate matrix filter UI options depending on available data and metadata,
as well as handling of user inputs to pass on appropriate plot objects.
}
\details{
See \code{\link{metadata}} for an example data object that can be passed into the parameter
\preformatted{metadata}. Since the filter UI is composed mainly of drop-down selections, one way to expand and improve it is to integrate
with an "add-on" or "augmenting" widget that provides an alternative and more intuitive or interesting
route of input for users. Foe example, in addition to a drop-down selection of geographical locations,
a map widget can be available to allow users to choose more visually. Because the widget may only need to appear for
certain metadata, the server function needs to return when the widget should be called (displayed).
}
