% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCtrl.R
\name{matrixCtrl}
\alias{matrixCtrl}
\title{Shiny module server functions to generate filter UI for interactive matrix}
\usage{
matrixCtrl(input, output, session, M = NULL, N = NULL, cdata = NULL,
  metadata = NULL, vkey = NULL, newdata = reactive({ }),
  widgetmod = NULL, widgetopt = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{M}{A data matrix, e.g. a correlation matrix, which must have row names.}

\item{N}{A matrix of the same dimensions as M, used as a filter layer, e.g. sample size.}

\item{cdata}{The data used for generating the matrix.}

\item{metadata}{Optional, a data.table with different types of metadata/annotation to be used as filters.
If not given, the only filter option will be the row names/index in M.}

\item{vkey}{The column in metadata that maps to row names in M.}

\item{newdata}{Optional, reactive data, e.g. from user upload, that can be merged with M.}

\item{widgetmod}{Optional, a widget extension module. See details.}
}
\value{
Reactive values in \code{state} object to be used by the associated plotting module.
Basically, this keeps track of all the matrices as well as selected metadata.
}
\description{
Update/populate matrix filter UI options depending on available data and metadata,
as well as handling of user inputs to pass on to appropriate plot objects.
}
\details{
What is available as filters relies on the underlying metadata.
See \code{\link{metadata}} for an example data object that can be passed into the parameter
\preformatted{metadata}. The base UI implements these filters as drop-down selections,
but the UI can also expand/integrate with an optional "add-on" or "augmenting" widget
to provide an alternate, perhaps more intuitive route of filter input.
For example, compared to a drop-down selection of geographical locations,
a map widget would provide a better selection interface. Not all types of metadata
can be integrated with a widget, and the module provides capability for only one widget.
The server function needs to return when the widget should be called (displayed).

The module returns an object representing the original data with user-applied filters,
i.e. the main input to be visualized with \code{\link{interactiveMatrix}}.
}
