% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meh.R
\name{meh}
\alias{meh}
\title{Modal Error Handling}
\usage{
meh(
  msg = getOption("DIVE.error.modal.msg"),
  error = NULL,
  footer = getOption("DIVE.error.modal.footer")
)
}
\arguments{
\item{msg}{User-facing message that can contain HTML for formatting.}

\item{error}{Forwarded error messsage to be shown depending on \code{dev_mode} option.}

\item{footer}{Bottom part of the modal dialog, which by default contains nothing unless
set with \code{base::options(DIVE.error.modal.footer = ui)}. See details for intended uses.}
}
\description{
Bring up a modal dialog for application errors
}
\details{
Module code will summon a modal dialog to display an error when execution fails,
something is seriously not allowed, etc.
This defaults to a generic user-facing message of
"Sorry, something went wrong. Try again later or let us know about the issue.",
which can be configured through \code{base::options(DIVE.error.modal.mg = altmessage)}.
When the \code{dev_mode} option is set, the original error will also be displayed.
The footer can optionally contain a link to where the user can file issues, etc.,
configured with \code{base::options(DIVE.error.modal.footer = ui)}.
}
