% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVUI.R
\name{multiV}
\alias{multiV}
\title{Shiny module server for presenting high dimensional data with other features in multi-column view}
\usage{
multiV(input, output, session, hdata, cdata = reactive({     NULL }),
  key = "ID", selected = reactive({     NULL }), slabel = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{hdata}{A numeric matrix of of the high dimensional data.}

\item{cdata}{Optional, a data.frame or data.table that can be joined to high-dimensional data. See details.}

\item{selected}{A vector used to subset the columns of hdata.}

\item{slabel}{Optional, a vector that can map column names of hdata to plot labels.}
}
\description{
This module attempts to integratively visualize one or more
high-dimensional gene/protein/methylation datasets with other phenotype or clinical features.
}
