% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVCtrl.R
\name{multiVCtrlServer}
\alias{multiVCtrlServer}
\title{Shiny module server for controlling multi-dataset views}
\usage{
multiVCtrlServer(
  id,
  hdlist,
  choices = DIVE::hdlistchoicesMake(hdlist),
  cdata,
  key = "ID",
  preselect = NULL,
  checkFun = NULL,
  informd = system.file("info/ht_upload.Rmd", package = "DIVE")
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{hdlist}{A list of matrices representing high dimensional datasets; the names are used for \code{choices}.}

\item{choices}{Selection choices are by default created from automatic parsing of `hdlist`. However, a manual list can be given,
which should be appropriate for passing to \code{shiny::\link[shiny]{selectizeInput}}.}

\item{cdata}{A \code{data.table} of characteristics data, commonly phenotype or clinical data.}

\item{key}{Name of column that contains IDs in \code{cdata} matching sample IDs in \code{hdlist} datasets. Defaults to "ID".
Note that column should already be of class character.}

\item{preselect}{Optional, pre-selected phenotype or clinical variables from \code{cdata}.
If is \code{NULL} (not recommended for most cases), the user can dynamically render as many datasets views as they can source.}

\item{checkFun}{Optional, a custom check function for an additional layer of checking/modifying uploaded data.
It should return a list containing message and result (result should be \code{NULL} when data fails checks).}

\item{informd}{Relative path to the Rmarkdown file, whose contents will be displayed in the modal.}
}
\value{
A reactive values list containing the data matrix
for the parameter \preformatted{hdata} of the \code{\link{multiVServer}} module,
as well as parameters for \code{\link{geneVServer}} and \code{\link{selectVServer}}.
}
\description{
Implement control hub logic that provides data and parameters for \code{\link{xVServer}},
\code{\link{geneVServer}} and \code{\link{selectVServer}}
}
\details{
The server logic handles sourcing of large expression datasets with three different methods:
\enumerate{
  \item Selecting from available pre-processed datasets.
  \item User-uploaded data.
}
The data in \code{cdata} is supposed to be a phenotype or clinical
feature that one usually tries to correlate with expression data and can be numeric or categorical.
The module handles upload of phenotype/clinical data,
using a mutable version of \code{cdata} that appends user uploaded data.
}
