% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVCtrl.R
\name{multiVCtrlUI}
\alias{multiVCtrlUI}
\title{Shiny module UI for controlling multi-dataset views}
\usage{
multiVCtrlUI(id, menu = TRUE, upload = TRUE, maxItems = 3L)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{menu}{Logical flag, whether to allow a menu for loading stored datasets.}

\item{upload}{Logical flag, whether to allow data upload.}

\item{maxItems}{Integer representing the max number of tracks that can be selected (displayed).}
}
\description{
UI for sourcing and selecting datasets
}
\details{
Controls data views geared towards bioinformatics data.
Multiple high dimensional datasets can be displayed by calling separate instances of
\code{\link{xVUI}} modules, with the contents typically laid out in row containers.
The controller UI has three different ways to source the datasets:
\enumerate{
  \item Selecting from available pre-processed datasets.
  \item User-uploaded data.
  \item A beta (least-supported) method of retrieving datasets from GEO.
}
For customization, it is possible make unavailable any of the three sourcing methods,
e.g. hide the GEO sourcing option by not displaying the UI.
}
