% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVCtrlUI.R
\name{multiVCtrlUI}
\alias{multiVCtrlUI}
\title{Shiny module UI for controlling multi-column view}
\usage{
multiVCtrlUI(id, menu = T, upload = T, GEO = T)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{menu}{Logical flag, should there be a menu for sourcing stored datasets? See details.}

\item{upload}{Logical flag, should users be able to upload data? See details.}

\item{GEO}{Logical flag, should users be able to pull data from GEO? (beta) See details.}
}
\description{
Shiny module UI for controlling multi-column view
}
\details{
Multiple high dimensional datasets can be displayed by calling separate instances of \code{\link{multiVUI}} modules,
with the contents typically laid out in separate a row containers. This "controller" UI establishes three different ways to
select and source the datasets:
\enumerate{
  \item A selection menu that contains datasets that are pre-processed and stored in memory.
  \item An upload method for users to upload their own datasets.
  \item A beta (and least-supported) method for retrieving datasets from GEO.
}
For convenience and customization, it is possible make unavailable any of the three sourcing methods,
e.g. hide the GEO method by not displaying the UI.
}
