% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVUI.R
\name{multiVUI}
\alias{multiVUI}
\title{Shiny app UI for multi-views}
\usage{
multiVUI(id, CSS = system.file("www/", "app.css", package = "DIVE"))
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{CSS}{Optional, location to an alternate CSS stylesheet to change the look and feel of the app.}
}
\description{
Assembles the UI of various module components, i.e. \code{\link{xVUI}}, \code{\link{geneVUI}},
into a working one-page application
}
\seealso{
Other multiVUI module functions: 
\code{\link{geneVUI}()},
\code{\link{multiVUIR}()},
\code{\link{multiV}()},
\code{\link{selectV}()}
}
\concept{multiVUI module functions}
