% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiV.R
\name{multiVUIR}
\alias{multiVUIR}
\title{Shiny app launcher for multi-view module}
\usage{
multiVUIR(ns, ...)
}
\arguments{
\item{ns}{Namespace of app module.}

\item{...}{Arguments passed to \code{\link{multiVServer}}.}
}
\description{
Shiny app launcher for multi-view module
}
\seealso{
Other multiVUI module functions: 
\code{\link{selectVServer}()}
}
\concept{multiVUI module functions}
