% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVApp.R, R/multiVUI.R
\name{multiVUIR}
\alias{multiVUIR}
\title{Shiny app launcher for multi-view module}
\usage{
multiVUIR(ns, ...)

multiVUIR(ns, ...)
}
\description{
Shiny app launcher for multi-view module

Shiny app launcher for multi-view module
}
\seealso{
Other multiVApp module functions: \code{\link{geneVUI}},
  \code{\link{geneV}}, \code{\link{multiVUI}},
  \code{\link{multiV}}, \code{\link{selectVUI}},
  \code{\link{selectV}}, \code{\link{subgroupVUI}},
  \code{\link{subgroupV}}

Other multiVApp module functions: \code{\link{geneVUI}},
  \code{\link{geneV}}, \code{\link{multiVUI}},
  \code{\link{multiV}}, \code{\link{selectVUI}},
  \code{\link{selectV}}, \code{\link{subgroupVUI}},
  \code{\link{subgroupV}}
}
\concept{multiVApp module functions}
