% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVUI.R
\name{multiVUIR}
\alias{multiVUIR}
\title{Shiny app launcher for multi-view module}
\usage{
multiVUIR(ns, ...)
}
\description{
Shiny app launcher for multi-view module
}
\seealso{
Other multiVUI module functions: \code{\link{geneVUI}},
  \code{\link{multiVUI}}
}
\concept{multiVUI module functions}
