% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{networkGraph}
\alias{networkGraph}
\title{Make custom network graph representing relationships between studies or data}
\usage{
networkGraph(adjm, edgemods = NULL, nodemods = NULL)
}
\arguments{
\item{adjm}{Adjacency matrix encoding relationships between source nodes.}

\item{edgemods}{An optional function of adjusting weight values in the adjacency matrix.
For aesthetic reasons, weights are scaled by sqrt. Set to \code{NULL} if this is unnecessary.}

\item{nodemods}{Optional, a \code{data.frame} with 'node' column containing node names matching names in adjacency matrix,
and 'size' and/or 'color' column(s) specifying aesthetics.}
}
\value{
An \code{\link[igraph]} with aesthetics.
}
\description{
This is a wrapper around several \code{\link[igraph]} methods to build a custom network graph
with annotations representing relationships between experimental studies or data.
In the original purpose, the "relationship" represented by the graph's weighted edges
is the number of shared samples between each study, but the function is made generic
with the idea that other relationships (e.g. correlation) can be represented as well.
Edges and weights are set by passing in any appropriate adjacency matrix.
After the graph is constructed, node sizes and colors are modified if a
\code{data.frame} containing columns with values for those aesthetics is also passed in.
While node size is typically used to indicate the amount of data the study generated,
color can be used for other annotations. Finally, the resulting object is by default
converted into an interactive \code{\link[visNetwork]} object so it can be part of the Shiny application.
}
