% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newDataset.R
\name{newDataset}
\alias{newDataset}
\title{Shiny module server function for handling new dataset input}
\usage{
newDataset(input, output, session, refkey, checkFun = NULL,
  infoRmd = NULL, appdata = NULL)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{refkey}{Optional, a named list containing name/label for creating a key-like column.}

\item{checkFun}{Optional, a custom check function for an additional layer of checking/modifying uploaded data.
It should return a list containing message and result (result should be \code{NULL} for unsuccessful data).}

\item{appdata}{Optional, the name (including extension) of one or more files stored in appdata that can be
mock-uploaded. See details.}
}
\value{
A reactive data.table of the processed upload.
}
\description{
This is essentially the \code{\link{dataUpload}} module, but the uploaded data
is modified to include a key-like column before it is returned as the reactive object.
Use \code{\link{dataUpload}} if there is no need for \preformatted{refkey},
which is used to track the dataset in the same manner as in \code{\link{refSubset}}.
}
