% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.dive_result}
\alias{plot.dive_result}
\title{Plot a DIVE result}
\usage{
\method{plot}{dive_result}(x, ...)
}
\arguments{
\item{x}{A \code{dive_result} object.}

\item{...}{Additional arguments passed to
\code{plot_dive_heatmap()}.}
}
\value{
The input object invisibly.
}
\description{
Plot a DIVE result
}
