% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dive_heatmap}
\alias{plot_dive_heatmap}
\title{Plot a DIVE heatmap}
\usage{
plot_dive_heatmap(x, show_title = TRUE)
}
\arguments{
\item{x}{A \code{dive_result} object.}

\item{show_title}{Logical. Add a plot title?}
}
\value{
A ggplot object.
}
\description{
Creates a heatmap of temporal association values across
alpha-diversity metrics and temporal lags.
}
