% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refSubset.R
\name{refSubsetInput}
\alias{refSubsetInput}
\title{Shiny module UI for selecting a subsetted dataset}
\usage{
refSubsetInput(id, name = id, label = "", subsets)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{name}{Optional, name for the main dataset.}

\item{label}{Optional, select input label.}

\item{subsets}{A named list of the available subset factors in the data.}
}
\value{
A \code{shiny::\link[shiny]{tagList}} for UI to subset a dataset.
}
\description{
Shiny module UI to subset a dataset through a selection menu.
}
