% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVUI.R
\name{selectV}
\alias{selectV}
\title{Shiny module server for selecting columns from a data table}
\usage{
selectV(
  input,
  output,
  session,
  data = reactive({     NULL }),
  key = "ID",
  label = HTML("<strong>Phenotype/Experimental variable(s)</strong>"),
  selected = reactive({     NULL }),
  countby = reactive({     NULL }),
  maxitems = 3
)
}
\arguments{
\item{input, output, session}{Standard \code{shiny} boilerplate.}

\item{data}{A reactive data.table. See details.}

\item{key}{A key column that is kept for every selected subset. Defaults to "ID".}

\item{label}{Label for variable select input.}

\item{selected}{Optional, initial selection.}

\item{countby}{The matrix representing a high-throughput dataset with sample IDs,
which are intersected to generate counts in the select options. See details.}

\item{maxitems}{Maximum number of items that can be selected. Defaults to 3.}
}
\value{
The subsetted data table.
}
\description{
Returns the selected columns.
}
\details{
This is a subcomponent that originally serves under \code{\link{multiVCtrl}}
but should be generic enough to be integrated with other modules.
When data is uploaded through \code{\link{multiVCtrl}}, it can be "high-throughput"
or "low-throughput" type data, the latter typically as phenotype or clinical variables.
This module handles changes in the table of the latter data type.
When the user adds new data columns to the table, the selection changes to include these new options.
When the user selects specific columns from the menu, a column-subsetted table is returned.
The most recent modification involves adding parenthesized counts for each variable option,
i.e. "phenotypeA (15)", calculated on intersections with another table.
}
\seealso{
Other multiVUI module functions: 
\code{\link{geneVUI}()},
\code{\link{multiVUIR}()},
\code{\link{multiVUI}()},
\code{\link{multiV}()}
}
\concept{multiVUI module functions}
