% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupV.R
\name{subgroupVServer}
\alias{subgroupVServer}
\title{Shiny module server for creating comparison plots between subgrouped data}
\usage{
subgroupVServer(id, cdata, hdlist)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{cdata}{A data.table of clinical, phenotype, or other experimental data used to define groups.}

\item{hdlist}{A list of named high-dimensional (high-throughput datasets)}
}
\description{
This primarily allows creation of volcano plots for highthroughput datasets based on user-defined contrast groups
}
\seealso{
Other multiVApp module functions: 
\code{\link{selectVUI}()},
\code{\link{subgroupVUI}()}
}
\concept{multiVApp module functions}
