% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupV.R
\name{subgroupVUI}
\alias{subgroupVUI}
\title{Shiny app UI for creating comparison plots between subgrouped data}
\usage{
subgroupVUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
UI for user to define groups for volcano plots
}
\seealso{
Other multiVApp module functions: 
\code{\link{selectVUI}()},
\code{\link{subgroupVServer}()}
}
\concept{multiVApp module functions}
