% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNetwork.R
\name{vBinColor}
\alias{vBinColor}
\title{Map to color based on binary data. Since this takes a logical vector, ncolorscale controls
the color mapping to colorT and colorF; use a value of 3 to assign colorT for a proportion = 0.5.}
\usage{
vBinColor(x, colorT = "#2196f3", colorF = "gray20", ncolorscale = 10)
}
\arguments{
\item{x}{A logical vector.}

\item{colorT}{Color for TRUE.}

\item{colorF}{Color for FALSE.}

\item{ncolorscale}{Number of colors to use.}
}
\description{
Map to color based on binary data. Since this takes a logical vector, ncolorscale controls
the color mapping to colorT and colorF; use a value of 3 to assign colorT for a proportion = 0.5.
}
