% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xV.R
\name{xVServer}
\alias{xVServer}
\title{Shiny module server for presenting high dimensional data with other features in multi-column view}
\usage{
xVServer(
  id,
  hdata,
  cdata = reactive({     NULL }),
  key = "ID",
  selected = reactive({     NULL }),
  slabel = NULL,
  width = reactive(1600)
)
}
\arguments{
\item{hdata}{A numeric matrix representing high dimensional data.}

\item{cdata}{Optional, a data.frame or data.table that can be joined to high-dimensional data. See details.}

\item{key}{Name of key column, should be and currently defaults to "ID".}

\item{selected}{A vector used to subset the columns of hdata.}

\item{slabel}{Optional, a vector that can map column names of hdata to plot labels.}

\item{width}{Custom width, defaults to 1600, only used when dendro subplots are involved.}

\item{input, output, session}{Standard \code{shiny} boilerplate.}
}
\description{
This module attempts to integratively visualize one or more
high-dimensional gene/protein/methylation datasets with other phenotype or clinical features.
}
