% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xV.R
\name{xVServer}
\alias{xVServer}
\title{Shiny module server for cross-view of multidimensional data}
\usage{
xVServer(
  id,
  hdata,
  cdata = reactive(NULL),
  key = "ID",
  selected = reactive(NULL),
  height = 500,
  returndata = NULL
)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}

\item{hdata}{A numeric matrix with row and column names for heatmap representation.}

\item{cdata}{Optional, a \code{data.table} of characteristics data. See details.}

\item{key}{Name of key column for \code{cdata}, currently defaults to "ID".}

\item{selected}{A reactive vector used to subset the features (cols) of \code{hdata}.}

\item{height}{Height for data plots.}

\item{returndata}{A reactive object that can be used to return the local version of data in this module.}
}
\description{
Integratively visualize a high-dimensional expression dataset with phenotype or clinical features
through clustering, cross-selection and cross-comparison
}
\details{
The module is geared towards bioinformatics uses;
the data objects here can be compared to \code{Biobase::\link[Biobase]{ExpressionSet}} class object
where \code{hdata} corresponds to a (transposed) \code{exprs} matrix in the \code{assayData} slot
and \code{cdata} corresponds to the data in the \code{phenoData} slot.

The capabilities implemented are clustering (of samples by features only) and data subsetting.
The data subsetting of expression data works through a "local" selection,
which is simply created from column names of the \code{hdata} data matrix,
or through an alternative/global selection, which is whatever is passed into \code{selected}.
For alternative/global selection to work, the data must have an alternative index; see \code{\link{xVExprs}}.
Local selection takes precedences over alternative/global selection.
}
