% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xV.R
\name{xVUI}
\alias{xVUI}
\title{Shiny module UI for presenting high dimensional data with other features in multi-column view}
\usage{
xVUI(id)
}
\arguments{
\item{id}{Character ID for specifying namespace, see \code{shiny::\link[shiny]{NS}}.}
}
\description{
Integratively visualize a high-dimensional expression dataset with phenotype or clinical features
}
\details{
The UI is based on "visual spreadsheet" or "line up" view.
The original use case presumes the high-dimensional data to be expression data, e.g.
gene or protein expression matrices. See \code{\link{xVServer}} for details.
}
